eBird Status and Trends color palettes for abundance data

Generate the color palettes used in the eBird Status and Trends relative abundance maps.

abundance_palette(
  n,
  season = c("weekly", "breeding", "nonbreeding", "migration", "prebreeding_migration",
    "postbreeding_migration", "year_round")
)

Arguments

n
integer; the number of colors to be in the palette.

season
character; the season to generate colors for or "weekly" to get the color palette used in the weekly abundance animations.

Value

A character vector of color hex codes.

Examples

# breeding season color palette
abundance_palette(10, season = "breeding")

#>  [1] "#D8B9B5" "#D6A7A1" "#D3958D" "#D28479" "#D07265" "#CE6151" "#CC503E"
#>  [8] "#BB4938" "#AA4233" "#993C2E"