Calculates bins (breaks) based on standard deviations of Box-Cox power-transformed data for mapping

Mapping species abundance across the full-annual cycle presents a challenge, in that patterns of concentration and dispersion in abundance change throughout the year, making it difficult to define color bins that suit all seasons and accurately reflect the detail of abundance predictions. To address this, we selected a method (described by Maciejewski et al. 2013) that first selects an optimal power (the Box-Cox method) for normalizing the data, then power transforms the entire year of non-zero data, constructs bins with the power-transformed data using standard-deviations, and then un-transforms the bins.

calc_bins(x)

Arguments

x
RasterStack or RasterBrick; original eBird Status and Trends product raster GeoTIFF with 52 bands, one for each week.

Value

A list with two elements: bins is a vector containing the break points of the bins and power is the optimal power used to transform data when calculating bins.

References

Ross Maciejewski, Avin Pattah, Sungahn Ko, Ryan Hafen, William S. Cleveland, David S. Ebert. Automated Box-Cox Transformations for Improved Visual Encoding. IEEE Transactions on Visualization and Computer Graphics, 19(1): 130-140, 2013.

Examples

# download and load example abundance data
sp_path <- ebirdst_download("example_data")

#> Data already exists, use force = TRUE to re-download.

abd <- load_raster("abundance_umean", sp_path)

# calculate bins for a single week for this example
year_bins <- calc_bins(abd)