Calculates spatial extent of non-zero data from Raster* object for plotting

After loading a RasterStack of results, there are lots of NA values and plots of individual raster layers will display at the full extent of the study extent. To show an ideal extent, this function trims away 0 and NA values and checks to make sure it returns a reasonable extent for plotting. The returned Extent object can then be used for plotting. To access a pre-calculated extent for the full annual cycle use load_fac_map_parameters().

calc_full_extent(x)

Arguments

x
Raster* object; either full RasterStack or subset.

Value

raster Extent object

Examples

# simple toy example
r <- raster::raster(nrow = 100, ncol = 100)
r[5025:5075] <- 1
raster::extent(r)

#> class      : Extent 
#> xmin       : -180 
#> xmax       : 180 
#> ymin       : -90 
#> ymax       : 90 

calc_full_extent(r)

#> class      : Extent 
#> xmin       : -93.6 
#> xmax       : 90 
#> ymin       : -1.8 
#> ymax       : 3.6 

if (FALSE) {
# download and load example abundance data
sp_path <- ebirdst_download("example_data")
abd <- load_raster("abundance", sp_path)

# calculate full extent
plot_extent <- calc_full_extent(abd)

# plot
raster::plot(abd[[1]], axes = FALSE, ext = plot_extent)
}