Converts cryptic cover class names to readable land cover names

Internal function that converts the cryptic predictor class names to readable land cover names.

convert_classes(x, by_cover_class = FALSE, pretty = FALSE)

Arguments

x

character; vector of land cover variable names to convert.

by_cover_class

logical; whether to replace FRAGSTATS cover class name with a name for the cover class as whole.

pretty

logical; whether to convert from capital case to title case.

Value

A vector of converted names.

Examples

predictors <- c("MCD12Q1_LCCS1_FS_C1_1500_ED", "astwbd_fs_c1_1500_pland",
                "elev_250m_sd")
ebirdst:::convert_classes(predictors, pretty = TRUE)
#> [1] "Barren ED"         "Ocean PLAND"       "Elevation 250m SD"