eBird Status and Trends predictors descriptions

Details on the eBird Status and Trends predictor variables or, for variables all derived from the same dataset, details on the dataset.

Usage

ebirdst_predictor_descriptions

Format

A data frame with 37 rows and 4 columns

dataset: dataset name.

predictor: predictor name or, if multiple variables are derived from this dataset, the pattern used to generate the names.

description: detailed description of the dataset or variable.

reference: a reference to consult for further information on the dataset.