Data frame of species with eBird Trends Data Products

This data frame contains a list of species for which eBird Trends Data Products are available including model-level information for the trend model. The two predictive performance metrics (rsquared and beta0) are based on a comparison of actual and estimated percent per year trends for a suite of simulations (see Fink et al. 2023 for further details).

ebirdst_trends_runs

Format

A data frame with 11 variables:

run_name

Combination of species code and run name.

species_code

The alphanumeric eBird species code uniquely identifying the species.

common_name

The English common name of the species.

season

Season that the trend was estimated for: breeding, nonbreeding, or resident.

region

The geographic region that the trend model was run for. Note that broadly distributed species (e.g. Barn Swallow) will only have trend estimates for a regional subset of their full range.

start_date

The start date (MM-DD format) of the season for which the trend was estimated.

end_date

The end date (MM-DD format) of the season for which the trend was estimated.

start_year

The start date of the trend time period.

end_year

The end date of the trend time period.

rsquared

R-squared value comparing the actual and estimated trends from the simulations.

beta0

The slope of a linear model fitting actual vs. estimated trends (actual ~ estimated) for the simulations. Positive values of beta0 indicate that the models are systematically underestimating the simulated trend for this species.