eBird Status and Trends weeks

eBird Status and Trends predictions are made for each of 52 weeks of the year. This data frame provides the boundaries of the weeks.

ebirdst_weeks

Format

A data frame with 52 rows and 5 columns:

week_number

Integer week number from 1-52.

date

Date of the midpoint of the week.

week_midpoint

Date of the midpoint of the week expressed as a fraction of the year, i.e. a number from 0-1.

week_start

Date of the start of the week expressed as a fraction of the year, i.e. a number from 0-1.

week_end

Date of the end of the week expressed as a fraction of the year, i.e. a number from 0-1.