Load eBird Status and Trends test data predictions

During eBird Status and Trends modeling, predictions are made for checklists in a test dataset that is not included in the model fitting process. This function loads these predictions in addition to the actual observed count on the associated checklist. These data are used by ebirdst_ppms() to get for calculating predictive performance metrics.

load_predictions(path, return_sf = FALSE)

Arguments

path
character; directory that the Status and Trends data for a given species was downloaded to. This path is returned by ebirdst_download() or get_species_path().

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

Data frame, or sf object if return_sf = TRUE, containing observed counts and model predictiosn for the test data.

Examples

# \donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
#> Data already exists, use force = TRUE to re-download.
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# test data
test_predictions <- load_predictions(path)
dplyr::glimpse(test_predictions)
#> Rows: 69,325
#> Columns: 17
#> $ sampling_event_id <int> 10052301, 10081039, 10143301, 10209225, 10216906, 10…
#> $ lon               <dbl> -86.00780, -84.27086, -86.19749, -82.53907, -84.6824…
#> $ lat               <dbl> 43.25270, 44.04061, 42.64337, 43.27222, 44.97813, 42…
#> $ year              <int> 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012…
#> $ date              <dbl> 0.1584699, 0.1748634, 0.1939891, 0.2131148, 0.215847…
#> $ obs               <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ pi_median         <dbl> 0.0012130352, 0.0000000000, 0.0056247482, 0.00103057…
#> $ pi_upper          <dbl> 1.624602e-03, 9.706170e-06, 5.906293e-03, 1.775320e-…
#> $ pi_lower          <dbl> 0.0010769033, 0.0000000000, 0.0052052081, 0.00017210…
#> $ mu_median         <dbl> 0.7783889, 0.0000000, 0.9310000, 0.7955000, 0.628666…
#> $ mu_upper          <dbl> 0.78925107, 0.05334441, 0.96202950, 0.81088107, 0.73…
#> $ mu_lower          <dbl> 0.76657136, 0.00000000, 0.87568918, 0.78384603, 0.33…
#> $ pi_mu_median      <dbl> 0.0009054828, 0.0000000000, 0.0051081508, 0.00071076…
#> $ pi_mu_upper       <dbl> 1.206035e-03, 3.977800e-06, 6.086650e-03, 8.993073e-…
#> $ pi_mu_lower       <dbl> 7.216887e-04, 0.000000e+00, 4.116163e-03, 4.944714e-…
#> $ pat               <dbl> 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.000000…
#> $ pi_es             <int> 100, 84, 100, 100, 72, 101, 100, 101, 100, 100, 100,…
# }