Load summary data for eBird Status and Trends stixels

eBird Status and Trends divides space and time into variably sized "stixels" within which individual base models are fit. The process of stixelization is performed many times and the prediction at any given point is the median of the predictions from all the stixels that that point falls in. This function loads summary statistics for each stixel, for example, the size of the stixels, the number of observations within each stixel, and a suite of predictive performance metrics (PPMs) for the model fit within that stixel.

load_stixels(path, ext, return_sf = FALSE)

Arguments

path

character; directory that the Status and Trends data for a given species was downloaded to. This path is returned by ebirdst_download() or get_species_path().

ext

ebirdst_extent object; the spatiotemporal extent to filter the data to. The spatial component of the extent object must be provided in unprojected, latitude-longitude coordinates.

return_sf

logical; whether to return an sf object of spatial points rather then the default data frame.

Value

Data frame, or sf object if return_sf = TRUE, containing stixel summary data. Data are organized with one stixel per row and each stixel identified by a unique stixel_id, the centroid of each stixel in space and time is specified by lat, lon, and date.

Examples

# \donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
#> Data already exists, use force = TRUE to re-download.
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# load stixel summary information
stixels <- load_stixels(path)
dplyr::glimpse(stixels)
#> Rows: 250
#> Columns: 96
#> $ stixel_id                    <chr> "1-242.1-MR232", "1-185.8-OR142", "1-73.2…
#> $ srd_n                        <int> 15788, 15252, 15196, 14976, 15540, 15120,…
#> $ longitude                    <dbl> -84.00009, -83.57263, -85.39118, -86.4636…
#> $ latitude                     <dbl> 44.91965, 42.65483, 42.74910, 42.10683, 4…
#> $ day_of_year                  <dbl> 255.886231, 199.246430, 88.654779, 59.274…
#> $ stixel_width                 <dbl> 2.57401, 3.15333, 3.16124, 3.12097, 3.001…
#> $ stixel_height                <dbl> 1.78523, 2.21155, 2.21155, 2.18490, 2.078…
#> $ stixel_area                  <int> 138591, 133886, 133394, 131463, 136414, 1…
#> $ start_day_of_year            <dbl> 242.8410, 186.0000, 73.8954, 45.0000, 158…
#> $ end_day_of_year              <dbl> 270.0000, 213.0000, 101.0000, 73.0000, 18…
#> $ train_n                      <int> 1881, 3782, 4206, 3823, 4361, 4720, 2504,…
#> $ positive_ob_n                <int> 339, 32, 456, 73, 821, 970, 151, 8, 61, 1…
#> $ stixel_prevalence            <dbl> 0.1802232855, 0.0084611317, 0.1084165478,…
#> $ mean_non_zero_count          <dbl> 1.383481, 1.656250, 1.609649, 1.136986, 1…
#> $ train_n_occ                  <int> 1881, 3782, 4206, 3823, 4361, 4720, 2504,…
#> $ n_pos_occ                    <int> 339, 32, 456, 73, 821, 970, 151, 0, 61, 0…
#> $ train_n_abund                <int> 864, 1304, 1606, 1342, 2097, 2186, 903, 0…
#> $ n_pos_abund                  <int> 654, 1255, 1350, 1254, 1640, 1788, 792, 0…
#> $ is_assumed_zero              <int> 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0,…
#> $ neg_grid_cells               <int> 1542, 3750, 3750, 3750, 3540, 3750, 2353,…
#> $ pos_grid_cells               <int> 339, 32, 456, 73, 821, 970, 151, 0, 61, 0…
#> $ stixel_n                     <int> 3953, 16188, 22882, 20241, 8436, 31585, 5…
#> $ st_pos_n                     <int> 384, 31, 463, 66, 909, 1210, 151, 8, 57, …
#> $ avg_spat_cov                 <dbl> 0.1017819, 0.4109253, 0.5539085, 0.410039…
#> $ mean_occ_pred                <dbl> 0.3848300, 0.1589455, 0.3015019, 0.345300…
#> $ occ_thresh                   <dbl> 0.26782996, 0.04589660, 0.26569212, 0.087…
#> $ test_binary_tp               <int> 16, 0, 19, 2, 38, 51, 4, NA, 1, NA, 1, 55…
#> $ test_binary_fn               <int> 12, 0, 8, 3, 15, 13, 6, NA, 4, NA, 3, 33,…
#> $ test_binary_fp               <int> 47, 11, 143, 36, 76, 290, 30, NA, 27, NA,…
#> $ test_binary_tn               <int> 249, 1243, 1478, 1484, 520, 1890, 416, NA…
#> $ test_binary_auc              <dbl> 0.7063224, NA, 0.8077433, 0.6881579, 0.79…
#> $ test_binary_pcc              <dbl> 0.8179012, 0.9912281, 0.9083738, 0.974426…
#> $ test_binary_kappa            <dbl> 0.263522885, 0.000000000, 0.177970693, 0.…
#> $ test_binary_sensitivity      <dbl> 0.5714286, NA, 0.7037037, 0.4000000, 0.71…
#> $ test_binary_specificity      <dbl> 0.8412162, 0.9912281, 0.9117829, 0.976315…
#> $ train_oob_binary_tp          <int> 12, 0, 4, 0, 43, 43, 0, NA, 0, NA, 0, 113…
#> $ train_oob_binary_fn          <int> 3, 0, 1, 0, 11, 9, 0, NA, 0, NA, 0, 27, 9…
#> $ train_oob_binary_fp          <int> 128, 51, 470, 145, 199, 1011, 79, NA, 82,…
#> $ train_oob_binary_tn          <int> 691, 4606, 6398, 4769, 1594, 8275, 1237, …
#> $ train_oob_binary_auc         <dbl> 0.8218559, NA, 0.8657833, NA, 0.8426546, …
#> $ train_oob_binary_pcc         <dbl> 0.8429257, 0.9890487, 0.9314710, 0.970492…
#> $ train_oob_binary_kappa       <dbl> 0.1264571840, 0.0000000000, 0.0152836854,…
#> $ train_oob_binary_sensitivity <dbl> 0.8000000, NA, 0.8000000, NA, 0.7962963, …
#> $ train_oob_binary_specificity <dbl> 0.8437118, 0.9890487, 0.9315667, 0.970492…
#> $ test_occ_auc                 <dbl> 0.6156915, NA, 0.6455650, NA, 0.7067175, …
#> $ test_occ_pcc                 <dbl> 0.6349206, NA, 0.6481481, NA, 0.7017544, …
#> $ test_occ_kappa               <dbl> 0.05602606, NA, 0.14229983, NA, 0.3200000…
#> $ test_occ_bernoulli_dev       <dbl> -1.319916e-01, NA, -9.467997e-01, NA, 3.0…
#> $ test_occ_sensitivity         <dbl> 0.3125000, NA, 0.6315789, NA, 0.5263158, …
#> $ test_occ_specificity         <dbl> 0.7446809, NA, 0.6503497, NA, 0.7894737, …
#> $ test_occ_mse                 <dbl> 0.2248076, NA, 0.2477184, NA, 0.2131396, …
#> $ test_occ_spearman            <dbl> 0.2086024, NA, 0.1667651, NA, 0.3632171, …
#> $ train_oob_occ_auc            <dbl> 0.7526042, NA, NA, NA, 0.7324997, 0.87316…
#> $ train_oob_occ_pcc            <dbl> 0.7642857, NA, NA, NA, 0.6652893, 0.52182…
#> $ train_oob_occ_kappa          <dbl> 0.19512195, NA, NA, NA, 0.22772043, 0.066…
#> $ train_oob_occ_bernoulli_dev  <dbl> -1.00669711, NA, NA, NA, -0.37395900, -3.…
#> $ train_oob_occ_sensitivity    <dbl> 0.5833333, NA, NA, NA, 0.6744186, 0.93023…
#> $ train_oob_occ_specificity    <dbl> 0.7812500, NA, NA, NA, 0.6633166, 0.50445…
#> $ train_oob_occ_mse            <dbl> 0.1958517, NA, NA, NA, 0.2247569, 0.30093…
#> $ train_oob_occ_spearman       <dbl> 0.24618858, NA, NA, NA, 0.30878392, 0.255…
#> $ test_abd_spearman            <dbl> 0.2632439, NA, NA, NA, 0.3612416, 0.34949…
#> $ test_abd_poisson_dev         <dbl> 0.15862264, NA, NA, NA, 0.12000672, 0.070…
#> $ test_log_abd_pearson         <dbl> 0.2999289, NA, NA, NA, 0.4048512, 0.46062…
#> $ test_abd_mad                 <dbl> 0.4983436, NA, NA, NA, 0.6913270, 0.60723…
#> $ test_cnt_spearman            <dbl> 0.5986653, NA, NA, NA, 0.4686897, 0.46768…
#> $ test_log_cnt_pearson         <dbl> 0.5467011, NA, NA, NA, 0.4575831, 0.54165…
#> $ test_cnt_mad                 <dbl> 0.5889206, NA, NA, NA, 1.0997304, 0.81463…
#> $ train_oob_abd_spearman       <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_abd_poisson_dev    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_log_abd_pearson    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_abd_mad            <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_cnt_spearman       <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_log_cnt_pearson    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ train_oob_cnt_mad            <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ total_effort_hrs             <dbl> 2340.806, 3864.727, 3484.124, 3121.110, 5…
#> $ total_effort_distance_km     <dbl> 2768.663, 5588.941, 5052.824, 3888.280, 6…
#> $ total_number_observers       <int> 2675, 5234, 6389, 5718, 7636, 7041, 3303,…
#> $ max_time                     <dbl> -3.831527, -6.122889, -4.557751, -2.26055…
#> $ max_effort_hrs               <dbl> 5.000000, 2.483348, 2.500000, 6.000000, 3…
#> $ max_effort_distance_km       <dbl> 9.880000, 10.000000, 5.427087, 4.587000, …
#> $ cds_u10                      <dbl> 0.08935402, 0.87455226, 1.66916824, 0.737…
#> $ cds_v10                      <dbl> 0.72378974, -0.01465062, -0.13360358, 0.4…
#> $ cds_d2m                      <dbl> 10.9128574, 16.2915275, 0.9571457, -4.981…
#> $ cds_t2m                      <dbl> 17.0909184, 24.4125343, 8.7728453, 0.4709…
#> $ cds_hcc                      <dbl> 0.00000000, 0.12219798, 0.14896312, 0.073…
#> $ cds_i10fg                    <dbl> 6.192102, 6.353190, 8.422730, 7.494266, 6…
#> $ cds_mcc                      <dbl> 0.000000000, 0.104725863, 0.082584347, 0.…
#> $ cds_lcc                      <dbl> 0.037172112, 0.041917812, 0.090488761, 0.…
#> $ cds_sf                       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ cds_rf                       <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ cds_slc                      <dbl> -7.8677823, -3.2616315, -13.5573852, -12.…
#> $ cds_msl                      <dbl> 102151.4, 101525.0, 101501.2, 102064.3, 1…
#> $ day_occ_slope                <dbl> 34.417428, 590.876540, 3.628831, 27.99156…
#> $ day_occ_mag                  <dbl> 0.0219799377, 0.0004921375, 0.1813758286,…
#> $ day_cnt_slope                <dbl> 37.704040, -11.962687, 7.713641, 19.92061…
#> $ day_cnt_mag                  <dbl> 0.03067960, 0.08688800, 0.08962059, 0.033…
# }