Stixel summary file loader

Internal function used by load_pis() and load_pds() to get the stixel summary information (from summary.txt).

load_summary(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing stixel summary information about each stixel centroid.

Examples

# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# stixel summaries
summaries <- ebirdst:::load_summary(sp_path)
dplyr::glimpse(summaries)

#> Observations: 1,136
#> Variables: 202
#> $ stixel_id                                  <chr> "92-156.3-334", "69-121.4-…
#> $ srd_n                                      <int> 52385, 190745, 46020, 4931…
#> $ lon                                        <dbl> -83.89498, -82.66239, -83.…
#> $ lat                                        <dbl> 41.87075, 43.16171, 43.713…
#> $ date                                       <dbl> 0.4702172, 0.3719356, 0.74…
#> $ stixel_width                               <dbl> 4.126120, 8.112887, 4.1225…
#> $ stixel_height                              <dbl> 4.150, 8.125, 4.125, 3.825…
#> $ stixel_area                                <dbl> 17.12340, 65.91721, 17.005…
#> $ train_n                                    <int> 1228, 5471, 788, 628, 3429…
#> $ positive_ob_n                              <int> 33, 1080, 186, 16, 920, 38…
#> $ stixel_prevalence                          <dbl> 0.02687296, 0.19740450, 0.…
#> $ mean_non_zero_count                        <dbl> 1.636364, 1.379630, 1.3548…
#> $ binary_kappa                               <dbl> 1.0000000, 0.8111115, 1.00…
#> $ binary_auc                                 <dbl> 1.0000000, 0.9736369, 1.00…
#> $ binary_deviance_model                      <dbl> 123.13160, -552.86970, 130…
#> $ binary_deviance_mean                       <dbl> -442.9989, -6407.3880, -98…
#> $ binary_deviance_explained                  <dbl> 1.2779500, 0.9137137, 1.13…
#> $ pois_deviance_model                        <dbl> 8.5736200, 357.8219000, 36…
#> $ pois_deviance_mean                         <dbl> 8.1763920, 797.3943000, 58…
#> $ posi_deviance_explained                    <dbl> -0.04858240, 0.55126100, 0…
#> $ total_effort_hrs                           <dbl> 1491.200, 8508.853, 1263.6…
#> $ total_effort_distance_km                   <dbl> 2197.452, 10541.140, 1451.…
#> $ total_number_observers                     <int> 1991, 13082, 1394, 893, 65…
#> $ train_elevation_mean                       <dbl> 240.4226, 247.6543, 214.36…
#> $ train_cov_mean_i_stationary                <dbl> 0.3729642, 0.3121916, 0.32…
#> $ train_cov_mean_year                        <dbl> 2012.814, 2012.791, 2013.1…
#> $ train_cov_mean_day                         <dbl> 0.4734746, 0.4438198, 0.52…
#> $ train_cov_mean_time                        <dbl> 11.99189, 11.61687, 11.858…
#> $ train_cov_mean_effort_hrs                  <dbl> 1.214332, 1.555265, 1.6036…
#> $ train_cov_mean_effort_distance_km          <dbl> 1.789456, 1.926731, 1.8424…
#> $ train_cov_mean_number_observers            <dbl> 1.621336, 2.391153, 1.7690…
#> $ train_cov_mean_score                       <dbl> 0.050703040, 0.039286970, …
#> $ train_cov_mean_elev                        <dbl> 240.4226, 247.6543, 214.36…
#> $ train_cov_mean_northness                   <dbl> -0.0228395800, -0.00507806…
#> $ train_cov_mean_eastness                    <dbl> 1.237670e-02, 4.515770e-03…
#> $ train_cov_mean_umd_fs_c1_1500_pland        <dbl> 2.101106, 2.653488, 3.6681…
#> $ train_cov_mean_umd_fs_c1_1500_lpi          <dbl> 0.01628803, 0.02065913, 0.…
#> $ train_cov_mean_umd_fs_c1_1500_pd           <dbl> 4.634704e-08, 5.748883e-08…
#> $ train_cov_mean_umd_fs_c1_1500_ed           <dbl> 0.0001413377, 0.0001779888…
#> $ train_cov_mean_umd_fs_c2_1500_pland        <dbl> 0.120026300, 0.109399800, …
#> $ train_cov_mean_umd_fs_c2_1500_lpi          <dbl> 1.071004e-03, 8.969164e-04…
#> $ train_cov_mean_umd_fs_c2_1500_pd           <dbl> 4.051683e-09, 4.161925e-09…
#> $ train_cov_mean_umd_fs_c2_1500_ed           <dbl> 9.144856e-06, 8.889915e-06…
#> $ train_cov_mean_umd_fs_c3_1500_pland        <dbl> 0.3622024, 0.4948272, 0.71…
#> $ train_cov_mean_umd_fs_c3_1500_lpi          <dbl> 0.0027689140, 0.0039043220…
#> $ train_cov_mean_umd_fs_c3_1500_pd           <dbl> 1.389056e-08, 1.707007e-08…
#> $ train_cov_mean_umd_fs_c3_1500_ed           <dbl> 2.944329e-05, 3.900649e-05…
#> $ train_cov_mean_umd_fs_c4_1500_pland        <dbl> 8.489645, 12.523650, 7.950…
#> $ train_cov_mean_umd_fs_c4_1500_lpi          <dbl> 0.07285088, 0.10926970, 0.…
#> $ train_cov_mean_umd_fs_c4_1500_pd           <dbl> 8.427459e-08, 1.034047e-07…
#> $ train_cov_mean_umd_fs_c4_1500_ed           <dbl> 0.0003922504, 0.0005129881…
#> $ train_cov_mean_umd_fs_c5_1500_pland        <dbl> 6.211084, 15.041510, 11.32…
#> $ train_cov_mean_umd_fs_c5_1500_lpi          <dbl> 0.04659559, 0.13070950, 0.…
#> $ train_cov_mean_umd_fs_c5_1500_pd           <dbl> 1.084965e-07, 1.339699e-07…
#> $ train_cov_mean_umd_fs_c5_1500_ed           <dbl> 0.0003807680, 0.0006355795…
#> $ train_cov_mean_umd_fs_c6_1500_pland        <dbl> 0.30246630, 0.23632130, 0.…
#> $ train_cov_mean_umd_fs_c6_1500_lpi          <dbl> 0.0025584070, 0.0020589910…
#> $ train_cov_mean_umd_fs_c6_1500_pd           <dbl> 1.000663e-08, 8.466733e-09…
#> $ train_cov_mean_umd_fs_c6_1500_ed           <dbl> 2.174519e-05, 1.814795e-05…
#> $ train_cov_mean_umd_fs_c7_1500_pland        <dbl> 0.7771703, 0.6225362, 0.90…
#> $ train_cov_mean_umd_fs_c7_1500_lpi          <dbl> 0.006113955, 0.005158098, …
#> $ train_cov_mean_umd_fs_c7_1500_pd           <dbl> 2.673380e-08, 2.051228e-08…
#> $ train_cov_mean_umd_fs_c7_1500_ed           <dbl> 5.983514e-05, 4.702418e-05…
#> $ train_cov_mean_umd_fs_c8_1500_pland        <dbl> 3.376524, 2.856963, 3.7043…
#> $ train_cov_mean_umd_fs_c8_1500_lpi          <dbl> 0.02398310, 0.02067695, 0.…
#> $ train_cov_mean_umd_fs_c8_1500_pd           <dbl> 8.983812e-08, 8.212200e-08…
#> $ train_cov_mean_umd_fs_c8_1500_ed           <dbl> 0.0002414848, 0.0002102256…
#> $ train_cov_mean_umd_fs_c9_1500_pland        <dbl> 0.5421495, 0.4478261, 0.57…
#> $ train_cov_mean_umd_fs_c9_1500_lpi          <dbl> 0.004387884, 0.003784125, …
#> $ train_cov_mean_umd_fs_c9_1500_pd           <dbl> 2.041326e-08, 1.780041e-08…
#> $ train_cov_mean_umd_fs_c9_1500_ed           <dbl> 4.342462e-05, 3.645424e-05…
#> $ train_cov_mean_umd_fs_c10_1500_pland       <dbl> 3.941525, 3.983672, 4.6895…
#> $ train_cov_mean_umd_fs_c10_1500_lpi         <dbl> 0.02670631, 0.02760121, 0.…
#> $ train_cov_mean_umd_fs_c10_1500_pd          <dbl> 1.091073e-07, 1.082762e-07…
#> $ train_cov_mean_umd_fs_c10_1500_ed          <dbl> 0.0002850011, 0.0002851080…
#> $ train_cov_mean_umd_fs_c12_1500_pland       <dbl> 55.616350, 42.128080, 37.9…
#> $ train_cov_mean_umd_fs_c12_1500_lpi         <dbl> 0.53328240, 0.39772660, 0.…
#> $ train_cov_mean_umd_fs_c12_1500_pd          <dbl> 1.495252e-07, 1.470460e-07…
#> $ train_cov_mean_umd_fs_c12_1500_ed          <dbl> 0.0013553400, 0.0011436810…
#> $ train_cov_mean_umd_fs_c13_1500_pland       <dbl> 13.529230, 11.235090, 17.6…
#> $ train_cov_mean_umd_fs_c13_1500_lpi         <dbl> 0.12980290, 0.10723440, 0.…
#> $ train_cov_mean_umd_fs_c13_1500_pd          <dbl> 4.904602e-08, 4.714339e-08…
#> $ train_cov_mean_umd_fs_c13_1500_ed          <dbl> 3.702720e-04, 3.211699e-04…
#> $ train_cov_mean_umd_fs_c16_1500_pland       <dbl> 0.17191460, 0.23387590, 0.…
#> $ train_cov_mean_umd_fs_c16_1500_lpi         <dbl> 0.0013147500, 0.0018849130…
#> $ train_cov_mean_umd_fs_c16_1500_pd          <dbl> 5.931286e-09, 7.799628e-09…
#> $ train_cov_mean_umd_fs_c16_1500_ed          <dbl> 1.394147e-05, 1.801734e-05…
#> $ train_cov_mean_modiswater_fs_c0_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c0_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c0_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c0_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c2_1500_pland <dbl> 5.521994, 9.631437, 12.489…
#> $ train_cov_mean_modiswater_fs_c2_1500_lpi   <dbl> 0.05375331, 0.09330728, 0.…
#> $ train_cov_mean_modiswater_fs_c2_1500_pd    <dbl> 1.954228e-08, 3.163536e-08…
#> $ train_cov_mean_modiswater_fs_c2_1500_ed    <dbl> 1.603030e-04, 2.721159e-04…
#> $ train_cov_mean_modiswater_fs_c3_1500_pland <dbl> 4.850032, 7.316790, 10.375…
#> $ train_cov_mean_modiswater_fs_c3_1500_lpi   <dbl> 0.04666992, 0.07109098, 0.…
#> $ train_cov_mean_modiswater_fs_c3_1500_pd    <dbl> 1.831215e-08, 2.685509e-08…
#> $ train_cov_mean_modiswater_fs_c3_1500_ed    <dbl> 1.323424e-04, 2.032644e-04…
#> $ train_cov_mean_modiswater_fs_c5_1500_pland <dbl> 0.00000000, 0.00000000, 0.…
#> $ train_cov_mean_modiswater_fs_c5_1500_lpi   <dbl> 0.0000000000, 0.0000000000…
#> $ train_cov_mean_modiswater_fs_c5_1500_pd    <dbl> 0.000000e+00, 0.000000e+00…
#> $ train_cov_mean_modiswater_fs_c5_1500_ed    <dbl> 0.000000e+00, 0.000000e+00…
#> $ train_cov_mean_modiswater_fs_c6_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c6_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c6_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c6_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c7_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c7_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c7_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_cov_mean_modiswater_fs_c7_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ train_covariate_entropy                    <dbl> 0.6348637, 0.6595938, 0.64…
#> $ srd_elevation_mean                         <dbl> 243.2308, 246.2400, 215.10…
#> $ srd_cov_mean_i_stationary                  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_year                          <int> 2016, 2016, 2016, 2016, 20…
#> $ srd_cov_mean_day                           <dbl> 0.5233032, 0.5197505, 0.51…
#> $ srd_cov_mean_time                          <dbl> 20.653060, 5.796735, 8.215…
#> $ srd_cov_mean_effort_hrs                    <int> 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ srd_cov_mean_effort_distance_km            <int> 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ srd_cov_mean_number_observers              <int> 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ srd_cov_mean_score                         <dbl> 1.85, 1.85, 1.85, 1.85, 1.…
#> $ srd_cov_mean_elev                          <dbl> 243.2308, 246.2400, 215.10…
#> $ srd_cov_mean_northness                     <dbl> 0.0130616500, -0.006531132…
#> $ srd_cov_mean_eastness                      <dbl> 0.0047699020, 0.0003106797…
#> $ srd_cov_mean_umd_fs_c1_1500_pland          <dbl> 0.7609260, 0.9053681, 0.96…
#> $ srd_cov_mean_umd_fs_c1_1500_lpi            <dbl> 0.006455941, 0.007458329, …
#> $ srd_cov_mean_umd_fs_c1_1500_pd             <dbl> 1.547616e-08, 2.077987e-08…
#> $ srd_cov_mean_umd_fs_c1_1500_ed             <dbl> 4.903044e-05, 6.116012e-05…
#> $ srd_cov_mean_umd_fs_c2_1500_pland          <dbl> 0.108968900, 0.057158920, …
#> $ srd_cov_mean_umd_fs_c2_1500_lpi            <dbl> 9.093994e-04, 5.089692e-04…
#> $ srd_cov_mean_umd_fs_c2_1500_pd             <dbl> 3.248388e-09, 1.980974e-09…
#> $ srd_cov_mean_umd_fs_c2_1500_ed             <dbl> 8.103070e-06, 4.447609e-06…
#> $ srd_cov_mean_umd_fs_c3_1500_pland          <dbl> 0.09544717, 0.27057650, 0.…
#> $ srd_cov_mean_umd_fs_c3_1500_lpi            <dbl> 0.0007953931, 0.0022099020…
#> $ srd_cov_mean_umd_fs_c3_1500_pd             <dbl> 3.520116e-09, 8.375042e-09…
#> $ srd_cov_mean_umd_fs_c3_1500_ed             <dbl> 7.588044e-06, 2.004724e-05…
#> $ srd_cov_mean_umd_fs_c4_1500_pland          <dbl> 4.422120, 8.854899, 6.2602…
#> $ srd_cov_mean_umd_fs_c4_1500_lpi            <dbl> 0.03702820, 0.07819049, 0.…
#> $ srd_cov_mean_umd_fs_c4_1500_pd             <dbl> 5.580310e-08, 7.357079e-08…
#> $ srd_cov_mean_umd_fs_c4_1500_ed             <dbl> 0.0002250090, 0.0003588859…
#> $ srd_cov_mean_umd_fs_c5_1500_pland          <dbl> 2.358871, 12.221960, 9.382…
#> $ srd_cov_mean_umd_fs_c5_1500_lpi            <dbl> 0.01813231, 0.11239260, 0.…
#> $ srd_cov_mean_umd_fs_c5_1500_pd             <dbl> 4.800944e-08, 7.889975e-08…
#> $ srd_cov_mean_umd_fs_c5_1500_ed             <dbl> 0.0001520470, 0.0004114557…
#> $ srd_cov_mean_umd_fs_c6_1500_pland          <dbl> 0.16703250, 0.15421990, 0.…
#> $ srd_cov_mean_umd_fs_c6_1500_lpi            <dbl> 0.0013919380, 0.0013011850…
#> $ srd_cov_mean_umd_fs_c6_1500_pd             <dbl> 5.990373e-09, 5.780102e-09…
#> $ srd_cov_mean_umd_fs_c6_1500_ed             <dbl> 1.309310e-05, 1.232129e-05…
#> $ srd_cov_mean_umd_fs_c7_1500_pland          <dbl> 0.4019386, 0.3005030, 0.33…
#> $ srd_cov_mean_umd_fs_c7_1500_lpi            <dbl> 0.0033406510, 0.0025317400…
#> $ srd_cov_mean_umd_fs_c7_1500_pd             <dbl> 1.419163e-08, 1.080377e-08…
#> $ srd_cov_mean_umd_fs_c7_1500_ed             <dbl> 3.148523e-05, 2.367134e-05…
#> $ srd_cov_mean_umd_fs_c8_1500_pland          <dbl> 1.7965280, 1.4387300, 1.68…
#> $ srd_cov_mean_umd_fs_c8_1500_lpi            <dbl> 0.013895520, 0.011088100, …
#> $ srd_cov_mean_umd_fs_c8_1500_pd             <dbl> 4.208083e-08, 3.885965e-08…
#> $ srd_cov_mean_umd_fs_c8_1500_ed             <dbl> 1.209051e-04, 1.019272e-04…
#> $ srd_cov_mean_umd_fs_c9_1500_pland          <dbl> 0.12858850, 0.16084590, 0.…
#> $ srd_cov_mean_umd_fs_c9_1500_lpi            <dbl> 0.0011824840, 0.0014264240…
#> $ srd_cov_mean_umd_fs_c9_1500_pd             <dbl> 5.483970e-09, 6.780765e-09…
#> $ srd_cov_mean_umd_fs_c9_1500_ed             <dbl> 1.072398e-05, 1.340569e-05…
#> $ srd_cov_mean_umd_fs_c10_1500_pland         <dbl> 2.1374860, 2.1434960, 2.00…
#> $ srd_cov_mean_umd_fs_c10_1500_lpi           <dbl> 0.015131030, 0.015392130, …
#> $ srd_cov_mean_umd_fs_c10_1500_pd            <dbl> 6.042248e-08, 6.144750e-08…
#> $ srd_cov_mean_umd_fs_c10_1500_ed            <dbl> 1.558582e-04, 1.568136e-04…
#> $ srd_cov_mean_umd_fs_c12_1500_pland         <dbl> 70.713630, 43.514610, 39.5…
#> $ srd_cov_mean_umd_fs_c12_1500_lpi           <dbl> 0.69611740, 0.42328910, 0.…
#> $ srd_cov_mean_umd_fs_c12_1500_pd            <dbl> 1.376304e-07, 1.107039e-07…
#> $ srd_cov_mean_umd_fs_c12_1500_ed            <dbl> 1.359977e-03, 9.601742e-04…
#> $ srd_cov_mean_umd_fs_c13_1500_pland         <dbl> 5.4921890, 3.3812590, 4.39…
#> $ srd_cov_mean_umd_fs_c13_1500_lpi           <dbl> 0.052776980, 0.032539080, …
#> $ srd_cov_mean_umd_fs_c13_1500_pd            <dbl> 2.438143e-08, 1.602079e-08…
#> $ srd_cov_mean_umd_fs_c13_1500_ed            <dbl> 1.555950e-04, 9.820257e-05…
#> $ srd_cov_mean_umd_fs_c16_1500_pland         <dbl> 0.020680220, 0.017693780, …
#> $ srd_cov_mean_umd_fs_c16_1500_lpi           <dbl> 1.908943e-04, 1.638313e-04…
#> $ srd_cov_mean_umd_fs_c16_1500_pd            <dbl> 8.398873e-10, 7.598256e-10…
#> $ srd_cov_mean_umd_fs_c16_1500_ed            <dbl> 1.739642e-06, 1.493013e-06…
#> $ srd_cov_mean_modiswater_fs_c0_1500_pland   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c0_1500_lpi     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c0_1500_pd      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c0_1500_ed      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c2_1500_pland   <dbl> 1.4020130, 2.9195460, 2.52…
#> $ srd_cov_mean_modiswater_fs_c2_1500_lpi     <dbl> 0.013712580, 0.028059570, …
#> $ srd_cov_mean_modiswater_fs_c2_1500_pd      <dbl> 5.558078e-09, 1.145166e-08…
#> $ srd_cov_mean_modiswater_fs_c2_1500_ed      <dbl> 4.021778e-05, 8.221318e-05…
#> $ srd_cov_mean_modiswater_fs_c3_1500_pland   <dbl> 11.431390, 14.911710, 16.2…
#> $ srd_cov_mean_modiswater_fs_c3_1500_lpi     <dbl> 0.11412300, 0.14879380, 0.…
#> $ srd_cov_mean_modiswater_fs_c3_1500_pd      <dbl> 1.687185e-08, 2.461292e-08…
#> $ srd_cov_mean_modiswater_fs_c3_1500_ed      <dbl> 1.758870e-04, 2.427986e-04…
#> $ srd_cov_mean_modiswater_fs_c5_1500_pland   <dbl> 0.000000, 11.764980, 18.77…
#> $ srd_cov_mean_modiswater_fs_c5_1500_lpi     <dbl> 0.00000000, 0.11762650, 0.…
#> $ srd_cov_mean_modiswater_fs_c5_1500_pd      <dbl> 0.000000e+00, 1.648211e-08…
#> $ srd_cov_mean_modiswater_fs_c5_1500_ed      <dbl> 0.000000e+00, 1.761630e-04…
#> $ srd_cov_mean_modiswater_fs_c6_1500_pland   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c6_1500_lpi     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c6_1500_pd      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c6_1500_ed      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c7_1500_pland   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c7_1500_lpi     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c7_1500_pd      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_cov_mean_modiswater_fs_c7_1500_ed      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ srd_covariate_entropy                      <dbl> 0.6220777, 0.6249278, 0.60…
#> $ max_time                                   <dbl> 20.653060, 5.796735, 8.215…