Test data loader

Internal function used by compute_ppms() to get the full test data set for calculating predictive performance metrics. This file contains the observed counts and all predictor variables used for modeling for each checklist in the test dataset.

load_test_data(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing test data, including checklist locations, observed counts, and predictor variables.

Examples

# \donttest{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# test data
test_data <- ebirdst:::load_test_data(sp_path)
dplyr::glimpse(test_data)

#> Rows: 42,557
#> Columns: 76
#> $ sampling_event_id                <int> 37115997, 40358704, 51150923, 302452…
#> $ observer_id                      <int> 213429, 391173, 1215261, 52415, 7748…
#> $ lon                              <dbl> -84.79610, -84.37254, -84.57054, -84…
#> $ lat                              <dbl> 41.70007, 41.70973, 41.71224, 41.717…
#> $ i_stationary                     <int> 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, …
#> $ year                             <int> 2017, 2017, 2019, 2016, 2018, 2018, …
#> $ day                              <int> 144, 311, 1, 166, 159, 202, 16, 5, 3…
#> $ solar_noon_diff                  <dbl> 0.12500000, -2.49166667, -3.41500000…
#> $ effort_hrs                       <dbl> 0.117, 0.250, 0.050, 0.033, 0.134, 0…
#> $ effort_distance_km               <dbl> 1.609, 0.000, 0.000, 1.609, 0.113, 0…
#> $ number_observers                 <int> 3, 1, 1, 1, 1, 3, 1, 1, 1, 1, 2, 1, …
#> $ score                            <dbl> 1.444020000, 0.089720110, -1.6133630…
#> $ eastness_median                  <dbl> 3.177390e-01, 1.261546e-02, 1.744153…
#> $ eastness_sd                      <dbl> 0.31585800, 0.06852572, 0.19681576, …
#> $ elev_median                      <dbl> 328.0, 257.5, 282.0, 239.5, 306.5, 3…
#> $ elev_sd                          <dbl> 11.0832886, 2.9001232, 5.2734736, 1.…
#> $ intertidal_fs_c1_1500_ed         <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ intertidal_fs_c1_1500_pland      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ island                           <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c1_1500_ed      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c1_1500_pland   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c2_1500_ed      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c2_1500_pland   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c11_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c11_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c12_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c12_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c13_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c13_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c14_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c14_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c15_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c15_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c16_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c16_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c21_1500_ed     <dbl> 1.199094, 0.000000, 0.000000, 0.0000…
#> $ mcd12q1_lccs1_fs_c21_1500_pland  <dbl> 2.777778, 0.000000, 0.000000, 0.0000…
#> $ mcd12q1_lccs1_fs_c22_1500_ed     <dbl> 9.592753, 0.000000, 5.138975, 0.0000…
#> $ mcd12q1_lccs1_fs_c22_1500_pland  <dbl> 16.666667, 0.000000, 11.904762, 0.00…
#> $ mcd12q1_lccs1_fs_c255_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c255_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c31_1500_ed     <dbl> 10.791847, 0.000000, 5.138975, 0.000…
#> $ mcd12q1_lccs1_fs_c31_1500_pland  <dbl> 80.55556, 100.00000, 88.09524, 100.0…
#> $ mcd12q1_lccs1_fs_c32_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c32_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c41_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c41_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c42_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c42_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c43_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs1_fs_c43_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs2_fs_c25_1500_ed     <dbl> 7.194565, 0.000000, 5.138975, 0.0000…
#> $ mcd12q1_lccs2_fs_c25_1500_pland  <dbl> 11.111111, 0.000000, 11.904762, 0.00…
#> $ mcd12q1_lccs2_fs_c35_1500_ed     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs2_fs_c35_1500_pland  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs2_fs_c36_1500_ed     <dbl> 13.190035, 0.000000, 7.194565, 0.000…
#> $ mcd12q1_lccs2_fs_c36_1500_pland  <dbl> 77.77778, 100.00000, 85.71429, 100.0…
#> $ mcd12q1_lccs3_fs_c27_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs3_fs_c27_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs3_fs_c50_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs3_fs_c50_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs3_fs_c51_1500_ed     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mcd12q1_lccs3_fs_c51_1500_pland  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mod44w_oic_fs_c1_1500_ed         <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mod44w_oic_fs_c1_1500_pland      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mod44w_oic_fs_c2_1500_ed         <dbl> 0.000000, 0.000000, 2.077032, 0.0000…
#> $ mod44w_oic_fs_c2_1500_pland      <dbl> 0.000000, 0.000000, 1.183432, 0.0000…
#> $ mod44w_oic_fs_c3_1500_ed         <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ mod44w_oic_fs_c3_1500_pland      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ northness_median                 <dbl> -0.173290759, -0.169144332, -0.16797…
#> $ northness_sd                     <dbl> 0.30411705, 0.16482453, 0.12880686, …
#> $ ntl_mean                         <dbl> 0.00000000, 0.07270143, 0.26184678, …
#> $ ntl_sd                           <dbl> 0.00000000, 0.16570451, 0.30558669, …
#> $ obs                              <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ date                             <dbl> 2017.393, 2017.850, 2019.003, 2016.4…
#> $ res                              <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, …

# }