Raw test data loader

Internal function used by load_test_data() to get the full test data for calculating predictive performance metrics. This file contains the observed counts and model predicted relative occurrence and abundance values.

load_test_data_raw(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing test data, including checklist locations, observed counts, and predicted mean and associated error for relative occurrence and abundance.

Examples

# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# test data
test_data <- ebirdst:::load_test_data_raw(sp_path)
dplyr::glimpse(test_data)

#> Observations: 32,344
#> Variables: 92
#> $ sampling_event_id           <int> 1517751, 1525641, 1539721, 1548582, 15822…
#> $ row_id                      <int> 9326, 9619, 10545, 10879, 11935, 12194, 1…
#> $ lon                         <dbl> -89.96339, -83.44579, -82.43158, -85.4863…
#> $ lat                         <dbl> 46.61525, 42.82401, 42.96513, 43.08301, 4…
#> $ dates                       <dbl> 2004.022, 2004.030, 2004.074, 2004.082, 2…
#> $ i_stationary                <int> 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0,…
#> $ year                        <int> 2004, 2004, 2004, 2004, 2004, 2004, 2004,…
#> $ day                         <int> 8, 11, 27, 30, 46, 47, 49, 64, 67, 76, 87…
#> $ time                        <dbl> 11.00, 8.00, 16.00, 8.25, 12.50, 16.00, 1…
#> $ effort_hrs                  <dbl> 2.50, 3.00, 0.50, 1.00, 1.50, 1.25, 0.50,…
#> $ effort_distance_km          <dbl> 0.000, 0.000, 0.000, 0.000, 0.000, 3.219,…
#> $ number_observers            <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ elev                        <dbl> 420.3423, 296.2309, 185.6930, 275.6766, 1…
#> $ northness                   <dbl> 0.682532635, -0.005460231, -0.078693440, …
#> $ eastness                    <dbl> -0.096234319, 0.330774645, -0.010951025, …
#> $ umd_fs_c1_1500_pland        <dbl> 0.000000, 0.000000, 0.000000, 4.761905, 0…
#> $ umd_fs_c1_1500_lpi          <dbl> 0.00000000, 0.00000000, 0.00000000, 0.047…
#> $ umd_fs_c1_1500_pd           <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ umd_fs_c1_1500_ed           <dbl> 0.0000000000, 0.0000000000, 0.0000000000,…
#> $ umd_fs_c2_1500_pland        <dbl> 0.000000, 4.761905, 0.000000, 0.000000, 0…
#> $ umd_fs_c2_1500_lpi          <dbl> 0.00000000, 0.02380952, 0.00000000, 0.000…
#> $ umd_fs_c2_1500_pd           <dbl> 0.000000e+00, 2.218361e-07, 0.000000e+00,…
#> $ umd_fs_c2_1500_ed           <dbl> 0.0000000000, 0.0004111180, 0.0000000000,…
#> $ umd_fs_c3_1500_pland        <dbl> 0.000000, 4.761905, 0.000000, 0.000000, 0…
#> $ umd_fs_c3_1500_lpi          <dbl> 0.00000000, 0.04761905, 0.00000000, 0.000…
#> $ umd_fs_c3_1500_pd           <dbl> 0.000000e+00, 1.109181e-07, 0.000000e+00,…
#> $ umd_fs_c3_1500_ed           <dbl> 0.0000000000, 0.0003083385, 0.0000000000,…
#> $ umd_fs_c4_1500_pland        <dbl> 42.857143, 14.285714, 0.000000, 23.809524…
#> $ umd_fs_c4_1500_lpi          <dbl> 0.40476190, 0.07142857, 0.00000000, 0.095…
#> $ umd_fs_c4_1500_pd           <dbl> 2.218361e-07, 3.327542e-07, 0.000000e+00,…
#> $ umd_fs_c4_1500_ed           <dbl> 0.0020555899, 0.0009250155, 0.0000000000,…
#> $ umd_fs_c5_1500_pland        <dbl> 54.761905, 7.142857, 0.000000, 9.523810, …
#> $ umd_fs_c5_1500_lpi          <dbl> 0.45238095, 0.04761905, 0.00000000, 0.095…
#> $ umd_fs_c5_1500_pd           <dbl> 2.218361e-07, 2.218361e-07, 0.000000e+00,…
#> $ umd_fs_c5_1500_ed           <dbl> 0.0022611489, 0.0005138975, 0.0000000000,…
#> $ umd_fs_c6_1500_pland        <dbl> 0.000000, 0.000000, 0.000000, 0.000000, 0…
#> $ umd_fs_c6_1500_lpi          <dbl> 0.00000000, 0.00000000, 0.00000000, 0.000…
#> $ umd_fs_c6_1500_pd           <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ umd_fs_c6_1500_ed           <dbl> 0.0000000000, 0.0000000000, 0.0000000000,…
#> $ umd_fs_c7_1500_pland        <dbl> 0.000000, 7.142857, 0.000000, 2.380952, 0…
#> $ umd_fs_c7_1500_lpi          <dbl> 0.00000000, 0.04761905, 0.00000000, 0.023…
#> $ umd_fs_c7_1500_pd           <dbl> 0.000000e+00, 2.218361e-07, 0.000000e+00,…
#> $ umd_fs_c7_1500_ed           <dbl> 0.0000000000, 0.0005138975, 0.0000000000,…
#> $ umd_fs_c8_1500_pland        <dbl> 2.380952, 9.523810, 0.000000, 7.142857, 0…
#> $ umd_fs_c8_1500_lpi          <dbl> 0.02380952, 0.04761905, 0.00000000, 0.047…
#> $ umd_fs_c8_1500_pd           <dbl> 1.109181e-07, 3.327542e-07, 0.000000e+00,…
#> $ umd_fs_c8_1500_ed           <dbl> 0.0002055590, 0.0008222360, 0.0000000000,…
#> $ umd_fs_c9_1500_pland        <dbl> 0.000000, 16.666667, 0.000000, 4.761905, …
#> $ umd_fs_c9_1500_lpi          <dbl> 0.00000000, 0.11904762, 0.00000000, 0.047…
#> $ umd_fs_c9_1500_pd           <dbl> 0.000000e+00, 3.327542e-07, 0.000000e+00,…
#> $ umd_fs_c9_1500_ed           <dbl> 0.0000000000, 0.0011305745, 0.0000000000,…
#> $ umd_fs_c10_1500_pland       <dbl> 0.000000, 2.380952, 0.000000, 4.761905, 0…
#> $ umd_fs_c10_1500_lpi         <dbl> 0.00000000, 0.02380952, 0.00000000, 0.023…
#> $ umd_fs_c10_1500_pd          <dbl> 0.000000e+00, 1.109181e-07, 0.000000e+00,…
#> $ umd_fs_c10_1500_ed          <dbl> 0.0000000000, 0.0002055590, 0.0000000000,…
#> $ umd_fs_c12_1500_pland       <dbl> 0.000000, 21.428571, 0.000000, 40.476190,…
#> $ umd_fs_c12_1500_lpi         <dbl> 0.00000000, 0.07142857, 0.00000000, 0.357…
#> $ umd_fs_c12_1500_pd          <dbl> 0.000000e+00, 4.436722e-07, 0.000000e+00,…
#> $ umd_fs_c12_1500_ed          <dbl> 0.0000000000, 0.0014389129, 0.0000000000,…
#> $ umd_fs_c13_1500_pland       <dbl> 0.000000, 11.904762, 100.000000, 2.380952…
#> $ umd_fs_c13_1500_lpi         <dbl> 0.00000000, 0.11904762, 1.00000000, 0.023…
#> $ umd_fs_c13_1500_pd          <dbl> 0.000000e+00, 1.109181e-07, 1.109181e-07,…
#> $ umd_fs_c13_1500_ed          <dbl> 0.0000000000, 0.0006166770, 0.0013361335,…
#> $ umd_fs_c16_1500_pland       <dbl> 0.000000, 0.000000, 0.000000, 0.000000, 0…
#> $ umd_fs_c16_1500_lpi         <dbl> 0.00000000, 0.00000000, 0.00000000, 0.000…
#> $ umd_fs_c16_1500_pd          <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ umd_fs_c16_1500_ed          <dbl> 0.0000000000, 0.0000000000, 0.0000000000,…
#> $ modiswater_fs_c0_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c0_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c0_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c0_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c2_1500_pland <dbl> 0.00000, 0.00000, 0.00000, 0.00000, 0.000…
#> $ modiswater_fs_c2_1500_lpi   <dbl> 0.0000000, 0.0000000, 0.0000000, 0.000000…
#> $ modiswater_fs_c2_1500_pd    <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ modiswater_fs_c2_1500_ed    <dbl> 0.0000000000, 0.0000000000, 0.0000000000,…
#> $ modiswater_fs_c3_1500_pland <dbl> 0.000000, 0.000000, 0.000000, 0.000000, 0…
#> $ modiswater_fs_c3_1500_lpi   <dbl> 0.00000000, 0.00000000, 0.00000000, 0.000…
#> $ modiswater_fs_c3_1500_pd    <dbl> 0.000000e+00, 0.000000e+00, 0.000000e+00,…
#> $ modiswater_fs_c3_1500_ed    <dbl> 0.0000000000, 0.0000000000, 0.0000000000,…
#> $ modiswater_fs_c5_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c5_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c5_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c5_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c6_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c6_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c6_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c6_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c7_1500_pland <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c7_1500_lpi   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c7_1500_pd    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ modiswater_fs_c7_1500_ed    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ y                           <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…