Test data predictions loader

Loads the model predictions for each checklist in the test dataset. Median, and upper and lower confidence intervals are provided for predicted occurrence, count, and relative abundance.

load_test_preds(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing median, and upper and lower confidence intervals are provided for predicted occurrence, count, and relativce abundance.

Examples

# \donttest{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# test data
test_preds <- load_test_preds(sp_path)
dplyr::glimpse(test_preds)

#> Observations: 42,557
#> Variables: 20
#> $ sampling_event_id <int> 19300692, 19778152, 21136092, 43067220, 31239121, 3…
#> $ lon               <dbl> -84.32520, -83.33005, -83.70724, -85.57344, -83.451…
#> $ lat               <dbl> 46.50468, 42.48074, 42.25520, 42.08974, 42.52874, 4…
#> $ date              <dbl> 0.57650270, 0.68579230, 0.98360660, 0.14207650, 0.6…
#> $ obs               <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, …
#> $ year              <int> 2014, 2014, 2014, 2018, 2016, 2017, 2016, 2018, 201…
#> $ pi_median         <dbl> 0.0111540500, 0.0217457100, 0.0242227200, 0.0069282…
#> $ pi_upper          <dbl> 0.0131257644, 0.0252766555, 0.0399402517, 0.0072969…
#> $ pi_lower          <dbl> 0.0097154097, 0.0190532351, 0.0196101869, 0.0067241…
#> $ pi_se             <dbl> 0.0012771849, 0.0023254358, 0.0080939734, 0.0002593…
#> $ mu_median         <dbl> 1.3819165, 0.9229570, 0.5007788, 0.9220833, 0.97640…
#> $ mu_upper          <dbl> 1.4236446, 0.9292306, 0.5486083, 0.9810156, 1.02780…
#> $ mu_lower          <dbl> 1.3418318, 0.9057661, 0.4640106, 0.8406052, 0.94876…
#> $ mu_se             <dbl> 0.032513986, 0.010291606, 0.029547008, 0.046075548,…
#> $ pi_mu_median      <dbl> 0.0157289289, 0.0205929501, 0.0136928204, 0.0062755…
#> $ pi_mu_upper       <dbl> 0.0195412812, 0.0231160369, 0.0154034033, 0.0072581…
#> $ pi_mu_lower       <dbl> 0.0132323599, 0.0170529123, 0.0110360559, 0.0058412…
#> $ pi_mu_se          <dbl> 0.0022193199, 0.0024380267, 0.0019564225, 0.0006456…
#> $ pat               <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.0…
#> $ pi_es             <int> 100, 100, 100, 100, 97, 87, 100, 98, 99, 98, 100, 8…

# }