Map PI and PD centroid locations

Creates a map showing the stixel centroid locations for PIs and/or PDs, with an optional spatiotemporal subset using an st_extent list.

map_centroids(pis, pds, st_extent = NA, plot_pis = TRUE,
  plot_pds = TRUE)

Arguments

pis
data.frame; predictor importance data from load_pis().

pds
data.frame; partial depenence data from load_pds().

st_extent
list; Optional spatiotemporal filter using st_extent list.

plot_pis
logical; Default is TRUE. Set to FALSE to hide the plotting of PI stixel centroid locations.

plot_pds
logical; Default is TRUE. Set to FALSE to hide the plotting of PD stixel centroid locations.

Value

Plot showing locations of PIs and/or PDs.

Examples

# NOT RUN {
# download and load example data
dl_path <- tempdir()
sp_path <- download_data("example_data", path = dl_path)
pis <- load_pis(sp_path)
pds <- load_pds(sp_path)

# define a spatiotemporal extent to plot
ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

map_centroids(pis = pis, pds = pds, st_extent = ne_extent)

# }