Parse data cube layer names into dates

label_raster_stack() labels the layers of a data cube with the associated week dates in the format of "wYYYY.MM.DD", because of constraints in the raster package. This function converts that character vector into an ISO compliant Date vector.

parse_raster_dates(x)

Arguments

x
Raster object; labeled Status and Trends data cube.

Value

Date vector.

Examples

# download and load example abundance data
sp_path <- ebirdst_download("example_data")
#> Data already exists, use force = TRUE to re-download.
abd <- load_raster(sp_path, "abundance")

# parse dates
parse_raster_dates(abd)
#>  [1] "2019-01-04" "2019-01-11" "2019-01-18" "2019-01-25" "2019-02-01"
#>  [6] "2019-02-08" "2019-02-15" "2019-02-22" "2019-03-01" "2019-03-08"
#> [11] "2019-03-15" "2019-03-22" "2019-03-29" "2019-04-05" "2019-04-12"
#> [16] "2019-04-19" "2019-04-26" "2019-05-03" "2019-05-10" "2019-05-17"
#> [21] "2019-05-24" "2019-05-31" "2019-06-07" "2019-06-14" "2019-06-21"
#> [26] "2019-06-28" "2019-07-06" "2019-07-13" "2019-07-20" "2019-07-27"
#> [31] "2019-08-03" "2019-08-10" "2019-08-17" "2019-08-24" "2019-08-31"
#> [36] "2019-09-07" "2019-09-14" "2019-09-21" "2019-09-28" "2019-10-05"
#> [41] "2019-10-12" "2019-10-19" "2019-10-26" "2019-11-02" "2019-11-09"
#> [46] "2019-11-16" "2019-11-23" "2019-11-30" "2019-12-07" "2019-12-14"
#> [51] "2019-12-21" "2019-12-28"