Generate stixel polygons from PI/PD data

All predictor importance and partial dependence data are provided at the stixel level. In these files, the stixel is defined based on a centroid, width, and height. This function uses this information to define polygons for each stixel and attaches them to the original data in the form of an sf object

stixelize(x)

# S3 method for data.frame
stixelize(x)

# S3 method for sf
stixelize(x)

Arguments

x
data.frame or sf object; PI or PD data loaded with load_pis() or load_pds(), or any other data frame with fields lon, lat, stixel_width, and stixel_hight.

Value

sf object with geometry column storing polygons representing the s tixels boundaries.

Methods (by class)

data.frame: PI or PD data

sf: PI or PD data as sf object

Examples

# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# load predictor importance
pis <- load_pis(sp_path)

stixelize(pis)

#> Simple feature collection with 849 features and 99 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -93.90373 ymin: 37.76792 xmax: -78.44666 ymax: 50.88339
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>        stixel_id  srd_n       lon      lat      date stixel_width stixel_height
#> 1   69-121.4-289 190745 -82.66239 43.16171 0.3719356     8.112887         8.125
#> 2    7-255.2-342  46020 -83.01648 43.71361 0.7401063     4.122559         4.125
#> 3   7-102.7-1150 142716 -84.55903 44.98568 0.3264353     5.963944         5.775
#> 4    10-68.4-680  41756 -86.21622 41.79261 0.2301514     4.196807         4.200
#> 5   20-348.7-330  40660 -85.36707 42.01676 0.9945355     3.971627         4.050
#> 6   44-264.6-303  43064 -84.86500 42.85970 0.7679523     4.054165         4.225
#> 7     1-181-1249  40564 -85.86240 42.68140 0.5376825     4.191180         4.200
#> 8       2-76-690  43184 -82.92381 42.43289 0.2506176     3.849929         3.825
#> 9       2-76-691  40580 -85.74791 43.50642 0.2509622     3.859651         3.825
#> 10 14-161.1-1151 137772 -89.04060 46.70954 0.4808628     7.043668         6.925
#>    stixel_area train_n positive_ob_n stixel_prevalence mean_non_zero_count
#> 1     65.91721    5471          1080        0.19740450            1.379630
#> 2     17.00556     788           186        0.23604060            1.354839
#> 3     34.44178    3429           920        0.26829980            1.604348
#> 4     17.62659    1752           389        0.22203200            2.272494
#> 5     16.08509    1069            96        0.08980355            1.072917
#> 6     17.12885    1469           358        0.24370320            1.254190
#> 7     17.60296     917            81        0.08833152            1.358025
#> 8     14.72598    2193           545        0.24851800            1.825688
#> 9     14.76317    1171           256        0.21861660            1.640625
#> 10    48.77740    2513           844        0.33585360            1.896919
#>    i_stationary         year         day        time effort_hrs
#> 1  0.0009782551 0.0753970200 0.093119370 0.074023190 0.33446210
#> 2  0.0009866625 0.0021844780 0.429846900 0.079498190 0.19003850
#> 3  0.0102888000 0.0755529600 0.563472600 0.167291600 0.57488890
#> 4  0.0040716770 0.0367294600 2.447557000 0.074426400 1.49242100
#> 5  0.0000000000 0.0007547261 0.006725923 0.008760425 0.04499456
#> 6  0.0087457020 0.0261479800 0.221674200 0.047343970 0.51563640
#> 7  0.0000000000 0.0117553100 0.038932490 0.024324970 0.02840611
#> 8  0.0007947534 0.0855764300 1.572839000 0.077741130 0.32078650
#> 9  0.0000000000 0.0295483100 2.528617000 0.065845020 0.33829840
#> 10 0.0017567630 0.0626012700 0.099072150 0.242724200 0.54061170
#>    effort_distance_km number_observers      score        elev   northness
#> 1         0.064391580     0.0088865010 0.10977060 0.101280800 0.058624140
#> 2         0.025708500     0.0055416540 0.25857680 0.038186440 0.008635983
#> 3         0.047441550     0.0145845300 0.27700610 0.051606640 0.050119220
#> 4         0.055355820     0.0106324500 0.14673300 0.024641730 0.043448080
#> 5         0.006688545     0.0003075676 0.03260055 0.006023087 0.016163430
#> 6         0.017106310     0.0047872150 0.18981830 0.039282690 0.033774600
#> 7         0.025159800     0.0023204540 0.02805700 0.132365400 0.009956901
#> 8         0.091529390     0.0059237920 0.15148180 0.027221590 0.026129700
#> 9         0.022476220     0.0021075910 0.13850190 0.025894900 0.015056400
#> 10        0.094723330     0.0056619270 0.39327290 0.429150600 0.138358900
#>        eastness umd_fs_c1_1500_pland umd_fs_c1_1500_lpi umd_fs_c1_1500_pd
#> 1  0.0498227900         0.0027234820       0.0043938350      0.0058794630
#> 2  0.0439463500         0.0003502530       0.0001266096      0.0004442638
#> 3  0.0807455400         0.0060644240       0.0043841150      0.0082682620
#> 4  0.0528379000         0.0003767716       0.0000000000      0.0015705880
#> 5  0.0037916700         0.0001454525       0.0005360913      0.0001232081
#> 6  0.0259320500         0.0011469630       0.0005356339      0.0031612130
#> 7  0.0009159445         0.0009700515       0.0000000000      0.0029727750
#> 8  0.0367949100         0.0030186030       0.0006602018      0.0018477210
#> 9  0.0407276700         0.0001568978       0.0005858426      0.0022897110
#> 10 0.1529488000         0.0017953200       0.0025331400      0.0123944400
#>    umd_fs_c1_1500_ed umd_fs_c2_1500_pland umd_fs_c2_1500_lpi umd_fs_c2_1500_pd
#> 1       1.090972e-02         6.536036e-05       1.385875e-04      1.026900e-03
#> 2       1.582699e-03         1.077418e-05       3.952960e-05      9.194824e-06
#> 3       7.763431e-03         5.425665e-05       5.851937e-04      6.861657e-04
#> 4       0.000000e+00         0.000000e+00       0.000000e+00      0.000000e+00
#> 5       1.034165e-04         1.635908e-04       5.725049e-03      1.998624e-04
#> 6       7.653757e-04         4.606220e-04       2.419310e-04      2.795407e-03
#> 7       5.057419e-06         0.000000e+00       1.295022e-10      0.000000e+00
#> 8       1.618128e-02         3.166677e-04       9.074745e-04      1.076650e-03
#> 9       6.114657e-04         7.054024e-06       7.327908e-05      3.181748e-04
#> 10      8.705668e-03         2.121127e-03       1.048519e-04      1.609226e-03
#>    umd_fs_c2_1500_ed umd_fs_c3_1500_pland umd_fs_c3_1500_lpi umd_fs_c3_1500_pd
#> 1       1.232356e-03         1.455139e-03       1.526986e-03      7.971544e-03
#> 2       2.129982e-04         0.000000e+00       3.034003e-05      0.000000e+00
#> 3       2.957774e-04         1.913617e-04       0.000000e+00      3.109620e-03
#> 4       0.000000e+00         4.595576e-05       1.206368e-04      0.000000e+00
#> 5       3.461667e-04         4.525723e-04       2.187336e-04      3.239615e-05
#> 6       7.645322e-04         8.663736e-05       5.011296e-04      4.581229e-04
#> 7       2.368322e-06         2.894102e-05       4.264299e-05      8.528113e-06
#> 8       4.978517e-04         9.545342e-04       1.052294e-03      1.098517e-02
#> 9       1.245962e-04         3.840394e-05       1.974018e-04      3.055801e-04
#> 10      0.000000e+00         1.718965e-03       8.333924e-04      1.390082e-02
#>    umd_fs_c3_1500_ed umd_fs_c4_1500_pland umd_fs_c4_1500_lpi umd_fs_c4_1500_pd
#> 1       8.491477e-04         0.0118347200       0.0322640700      0.0165489900
#> 2       2.807541e-02         0.0003564676       0.0015225970      0.0002984091
#> 3       1.399894e-03         0.0135092800       0.0180964800      0.0170723000
#> 4       8.711829e-04         0.0389391300       0.0159553100      0.0000000000
#> 5       2.849426e-04         0.0030635470       0.0003596819      0.0006753856
#> 6       3.872048e-04         0.0033715010       0.0040507510      0.0034322360
#> 7       1.531512e-05         0.0003196179       0.0095018490      0.0006281342
#> 8       1.097212e-03         0.0000000000       0.0029066850      0.0036290150
#> 9       1.989313e-05         0.0118694200       0.0014086050      0.0000000000
#> 10      2.397269e-03         0.0227168000       0.0399350800      0.0257360200
#>    umd_fs_c4_1500_ed umd_fs_c5_1500_pland umd_fs_c5_1500_lpi umd_fs_c5_1500_pd
#> 1       0.0099758480         0.0643741700       0.0072241630      0.0094194450
#> 2       0.0046034140         0.0011754040       0.0015831200      0.0002397078
#> 3       0.0194793600         0.0263829500       0.0433506700      0.0531123600
#> 4       0.0075254340         0.0122201100       0.0002632751      0.0000000000
#> 5       0.0002279311         0.0026361000       0.0004159598      0.0000000000
#> 6       0.0051245310         0.0008610701       0.0011230750      0.0197389700
#> 7       0.0010909900         0.0628251900       0.0635573300      0.0020898600
#> 8       0.0041386510         0.0025750870       0.0075747480      0.0041839780
#> 9       0.0070110180         0.0015117300       0.0029289940      0.0130602500
#> 10      0.0221969300         0.0396101900       0.0292842000      0.0312701700
#>    umd_fs_c5_1500_ed umd_fs_c6_1500_pland umd_fs_c6_1500_lpi umd_fs_c6_1500_pd
#> 1        0.007366266         5.233432e-05       1.380878e-04      5.841338e-03
#> 2        0.005266960         5.087715e-04       6.911137e-05      4.114848e-04
#> 3        0.025607830         1.954680e-04       1.408856e-03      3.875513e-03
#> 4        0.003939228         6.479944e-05       0.000000e+00      1.120315e-02
#> 5        0.001174661         1.379272e-04       8.474243e-04      2.499540e-03
#> 6        0.001359950         0.000000e+00       1.653208e-03      0.000000e+00
#> 7        0.001586942         6.777296e-06       0.000000e+00      2.346564e-06
#> 8        0.012418150         4.663898e-04       1.730161e-03      3.672262e-03
#> 9        0.004332934         1.093363e-03       1.679315e-04      0.000000e+00
#> 10       0.042510870         7.813094e-05       1.911201e-03      1.172361e-03
#>    umd_fs_c6_1500_ed umd_fs_c7_1500_pland umd_fs_c7_1500_lpi umd_fs_c7_1500_pd
#> 1       6.391328e-04         5.640498e-04       2.633142e-04      5.600501e-03
#> 2       1.803258e-03         2.755346e-04       1.431500e-04      4.869038e-04
#> 3       7.496521e-04         1.691197e-03       0.000000e+00      2.321885e-03
#> 4       0.000000e+00         3.858310e-05       3.448250e-04      6.560645e-04
#> 5       4.439042e-04         0.000000e+00       1.136329e-05      1.200172e-04
#> 6       1.296503e-03         5.652319e-05       3.417103e-04      1.953913e-03
#> 7       2.330573e-07         9.284303e-06       5.518037e-06      1.421568e-05
#> 8       1.202988e-03         1.768961e-04       6.690362e-04      2.805375e-04
#> 9       5.898564e-05         0.000000e+00       0.000000e+00      5.057841e-04
#> 10      4.098645e-04         6.142973e-04       5.773952e-04      3.394403e-03
#>    umd_fs_c7_1500_ed umd_fs_c8_1500_pland umd_fs_c8_1500_lpi umd_fs_c8_1500_pd
#> 1       1.294010e-03         2.472845e-03       1.000584e-02      0.0182494800
#> 2       0.000000e+00         6.402638e-04       1.630647e-03      0.0000000000
#> 3       2.442502e-03         4.805627e-03       3.913934e-03      0.0132522100
#> 4       1.270205e-03         5.550385e-03       1.201889e-03      0.0000000000
#> 5       0.000000e+00         9.689634e-05       5.511230e-05      0.0001275135
#> 6       7.240330e-04         2.874408e-03       2.599331e-03      0.0041826080
#> 7       1.275092e-05         0.000000e+00       3.419007e-05      0.0000000000
#> 8       7.285159e-04         1.159885e-03       3.923804e-03      0.0025455950
#> 9       8.295934e-05         0.000000e+00       2.302469e-03      0.0027952730
#> 10      2.398857e-03         2.270147e-03       1.535899e-02      0.0075878610
#>    umd_fs_c8_1500_ed umd_fs_c9_1500_pland umd_fs_c9_1500_lpi umd_fs_c9_1500_pd
#> 1       5.020096e-03         1.312570e-02       1.047461e-03      9.112189e-04
#> 2       9.395646e-03         5.082500e-05       1.792404e-05      4.504748e-05
#> 3       4.112604e-03         1.924896e-03       1.402168e-03      3.578811e-03
#> 4       2.974373e-03         1.124381e-02       3.329991e-05      2.852331e-03
#> 5       4.828050e-05         3.649386e-04       4.504763e-04      1.001990e-04
#> 6       5.051250e-03         1.660827e-03       1.152216e-03      6.901364e-04
#> 7       3.718155e-06         0.000000e+00       0.000000e+00      0.000000e+00
#> 8       3.709961e-04         0.000000e+00       1.424751e-03      8.798311e-05
#> 9       2.816046e-04         9.378029e-05       3.585386e-04      8.247225e-04
#> 10      8.187648e-03         1.543305e-03       1.677419e-03      1.318444e-03
#>    umd_fs_c9_1500_ed umd_fs_c10_1500_pland umd_fs_c10_1500_lpi
#> 1       7.525897e-03          5.077317e-03        0.0056275280
#> 2       6.439477e-06          0.000000e+00        0.0000000000
#> 3       2.614111e-03          4.310695e-03        0.0059748070
#> 4       3.470937e-04          6.459465e-03        0.0055587130
#> 5       1.502924e-04          3.654462e-04        0.0008241602
#> 6       3.474777e-04          3.322055e-03        0.0016373330
#> 7       0.000000e+00          4.802601e-05        0.0008022391
#> 8       9.476459e-04          3.450309e-03        0.0045647170
#> 9       2.162118e-05          6.876998e-04        0.0000000000
#> 10      7.325498e-03          1.437191e-03        0.0081857350
#>    umd_fs_c10_1500_pd umd_fs_c10_1500_ed umd_fs_c12_1500_pland
#> 1        0.0054368170       0.0057728050          0.0000000000
#> 2        0.0002489004       0.0009617416          0.0014104860
#> 3        0.0131386800       0.0049837320          0.0004563543
#> 4        0.0020258800       0.0044471880          0.0198884900
#> 5        0.0010591500       0.0014818100          0.0007562517
#> 6        0.0035058750       0.0041458280          0.0119007000
#> 7        0.0010976530       0.0020050760          0.0023693210
#> 8        0.0040236310       0.0034751460          0.0147784100
#> 9        0.0086990240       0.0016580520          0.0017732110
#> 10       0.0046592460       0.0071767080          0.0126661100
#>    umd_fs_c12_1500_lpi umd_fs_c12_1500_pd umd_fs_c12_1500_ed
#> 1         0.0130861700       0.0139291800        0.004212341
#> 2         0.0000000000       0.0046999840        0.005488440
#> 3         0.0063664180       0.0151762200        0.031009010
#> 4         0.0152740600       0.0117102800        0.018385990
#> 5         0.0022990800       0.0006057489        0.001008587
#> 6         0.0205046500       0.0164028700        0.011252710
#> 7         0.0007324885       0.0017421320        0.003357973
#> 8         0.0186357900       0.0158348000        0.027817910
#> 9         0.0003935473       0.0048519320        0.005652433
#> 10        0.0201326300       0.0381519400        0.024380870
#>    umd_fs_c13_1500_pland umd_fs_c13_1500_lpi umd_fs_c13_1500_pd
#> 1           3.146028e-03        1.548292e-02       6.761831e-03
#> 2           0.000000e+00        1.775985e-03       1.018786e-03
#> 3           1.256673e-02        9.735008e-03       2.293037e-03
#> 4           4.098051e-02        1.261063e-02       1.010415e-02
#> 5           3.747120e-05        0.000000e+00       4.095015e-03
#> 6           2.225345e-02        1.662034e-02       5.295310e-03
#> 7           1.481655e-05        1.623951e-05       5.204746e-07
#> 8           9.872144e-03        4.312237e-02       7.012191e-03
#> 9           3.904330e-03        1.780807e-01       7.079755e-03
#> 10          4.836227e-04        9.201509e-04       1.488433e-03
#>    umd_fs_c13_1500_ed umd_fs_c16_1500_pland umd_fs_c16_1500_lpi
#> 1        0.0069343890          5.989389e-04        2.881679e-04
#> 2        0.0011464460          5.475057e-05        2.568530e-05
#> 3        0.0119386500          2.225828e-04        4.455246e-04
#> 4        0.0104906400          1.863171e-04        2.164185e-04
#> 5        0.0002266771          3.546646e-04        5.344158e-04
#> 6        0.0059594500          3.526997e-04        5.015967e-05
#> 7        0.0000226393          1.060825e-05        1.316949e-05
#> 8        0.0144625500          5.724629e-04        1.358344e-04
#> 9        0.0067691490          1.742886e-04        3.205752e-05
#> 10       0.0013423870          2.211702e-03        1.160202e-03
#>    umd_fs_c16_1500_pd umd_fs_c16_1500_ed modiswater_fs_c0_1500_pland
#> 1        3.826464e-03       9.059628e-04                           0
#> 2        2.000764e-05       1.617079e-03                           0
#> 3        8.183452e-04       3.279005e-03                           0
#> 4        3.619752e-04       3.324844e-05                           0
#> 5        1.948389e-03       2.763306e-04                           0
#> 6        0.000000e+00       0.000000e+00                           0
#> 7        5.519794e-05       1.291217e-06                           0
#> 8        2.238055e-03       1.412964e-03                           0
#> 9        2.735311e-05       0.000000e+00                           0
#> 10       0.000000e+00       2.423434e-03                           0
#>    modiswater_fs_c0_1500_lpi modiswater_fs_c0_1500_pd modiswater_fs_c0_1500_ed
#> 1                          0                        0                        0
#> 2                          0                        0                        0
#> 3                          0                        0                        0
#> 4                          0                        0                        0
#> 5                          0                        0                        0
#> 6                          0                        0                        0
#> 7                          0                        0                        0
#> 8                          0                        0                        0
#> 9                          0                        0                        0
#> 10                         0                        0                        0
#>    modiswater_fs_c2_1500_pland modiswater_fs_c2_1500_lpi
#> 1                 1.067327e-03              2.581512e-03
#> 2                 1.905170e-02              0.000000e+00
#> 3                 5.693548e-03              8.845807e-03
#> 4                 0.000000e+00              0.000000e+00
#> 5                 1.495959e-05              0.000000e+00
#> 6                 2.764553e-03              5.532147e-03
#> 7                 9.599133e-07              4.296279e-06
#> 8                 0.000000e+00              0.000000e+00
#> 9                 6.526358e-04              3.118275e-06
#> 10                1.952885e-02              1.852254e-02
#>    modiswater_fs_c2_1500_pd modiswater_fs_c2_1500_ed
#> 1              0.000000e+00             4.556112e-03
#> 2              5.322000e-05             0.000000e+00
#> 3              6.550407e-03             9.315848e-03
#> 4              1.550674e-02             1.442754e-03
#> 5              2.865152e-05             1.672565e-06
#> 6              2.095630e-03             0.000000e+00
#> 7              0.000000e+00             2.975016e-06
#> 8              3.848252e-04             0.000000e+00
#> 9              0.000000e+00             0.000000e+00
#> 10             3.123477e-03             2.098731e-03
#>    modiswater_fs_c3_1500_pland modiswater_fs_c3_1500_lpi
#> 1                 4.066281e-03              3.121652e-03
#> 2                 9.150461e-04              2.019311e-03
#> 3                 6.144256e-03              1.295595e-02
#> 4                 0.000000e+00              7.203367e-02
#> 5                 1.127380e-04              8.890402e-03
#> 6                 1.002220e-03              1.348484e-04
#> 7                 8.165212e-06              1.111249e-06
#> 8                 1.270143e-03              2.783997e-03
#> 9                 1.656058e-04              5.654289e-04
#> 10                2.266942e-02              5.855875e-04
#>    modiswater_fs_c3_1500_pd modiswater_fs_c3_1500_ed
#> 1              7.289954e-04             4.689107e-03
#> 2              3.139606e-04             2.628940e-03
#> 3              7.743777e-03             9.265749e-03
#> 4              1.146467e-03             6.863937e-03
#> 5              2.665627e-04             2.400638e-05
#> 6              1.627095e-04             7.416909e-04
#> 7              8.293932e-07             4.544068e-06
#> 8              4.701189e-03             9.542242e-03
#> 9              4.057823e-03             1.067300e-03
#> 10             4.813483e-04             8.116956e-04
#>    modiswater_fs_c5_1500_pland modiswater_fs_c5_1500_lpi
#> 1                 0.0030293660              5.312359e-02
#> 2                 0.0000000000              0.000000e+00
#> 3                 0.0056210160              2.296938e-03
#> 4                 0.0005640737              2.279971e-04
#> 5                 0.0000000000              0.000000e+00
#> 6                 0.0040434390              3.756638e-03
#> 7                 0.0030735310              3.953502e-05
#> 8                 0.0000000000              0.000000e+00
#> 9                 0.0021510310              3.260836e-04
#> 10                0.0004609059              2.169040e-04
#>    modiswater_fs_c5_1500_pd modiswater_fs_c5_1500_ed
#> 1              7.567552e-03             7.557046e-04
#> 2              0.000000e+00             0.000000e+00
#> 3              8.548060e-03             1.216775e-03
#> 4              7.547949e-05             4.408866e-05
#> 5              0.000000e+00             0.000000e+00
#> 6              6.677018e-03             1.302557e-03
#> 7              1.684493e-05             7.273663e-07
#> 8              0.000000e+00             0.000000e+00
#> 9              4.722435e-05             8.709963e-05
#> 10             3.561498e-03             3.036658e-03
#>    modiswater_fs_c6_1500_pland modiswater_fs_c6_1500_lpi
#> 1                            0                         0
#> 2                            0                         0
#> 3                            0                         0
#> 4                            0                         0
#> 5                            0                         0
#> 6                            0                         0
#> 7                            0                         0
#> 8                            0                         0
#> 9                            0                         0
#> 10                           0                         0
#>    modiswater_fs_c6_1500_pd modiswater_fs_c6_1500_ed
#> 1                         0                        0
#> 2                         0                        0
#> 3                         0                        0
#> 4                         0                        0
#> 5                         0                        0
#> 6                         0                        0
#> 7                         0                        0
#> 8                         0                        0
#> 9                         0                        0
#> 10                        0                        0
#>    modiswater_fs_c7_1500_pland modiswater_fs_c7_1500_lpi
#> 1                            0                         0
#> 2                            0                         0
#> 3                            0                         0
#> 4                            0                         0
#> 5                            0                         0
#> 6                            0                         0
#> 7                            0                         0
#> 8                            0                         0
#> 9                            0                         0
#> 10                           0                         0
#>    modiswater_fs_c7_1500_pd modiswater_fs_c7_1500_ed
#> 1                         0                        0
#> 2                         0                        0
#> 3                         0                        0
#> 4                         0                        0
#> 5                         0                        0
#> 6                         0                        0
#> 7                         0                        0
#> 8                         0                        0
#> 9                         0                        0
#> 10                        0                        0
#>                          geometry
#> 1  POLYGON ((-86.71883 39.0992...
#> 2  POLYGON ((-85.07776 41.6511...
#> 3  POLYGON ((-87.541 42.09818,...
#> 4  POLYGON ((-88.31462 39.6926...
#> 5  POLYGON ((-87.35288 39.9917...
#> 6  POLYGON ((-86.89208 40.7472...
#> 7  POLYGON ((-87.95799 40.5814...
#> 8  POLYGON ((-84.84877 40.5203...
#> 9  POLYGON ((-87.67774 41.5939...
#> 10 POLYGON ((-92.56243 43.2470...

# also works on sf objects
pis_sf <- sf::st_as_sf(pis, coords = c("lon", "lat"), crs = 4326)
stixelize(pis_sf)

#> Simple feature collection with 849 features and 99 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -93.90373 ymin: 37.76792 xmax: -78.44666 ymax: 50.88339
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>        stixel_id  srd_n      date stixel_width stixel_height stixel_area
#> 1   69-121.4-289 190745 0.3719356     8.112887         8.125    65.91721
#> 2    7-255.2-342  46020 0.7401063     4.122559         4.125    17.00556
#> 3   7-102.7-1150 142716 0.3264353     5.963944         5.775    34.44178
#> 4    10-68.4-680  41756 0.2301514     4.196807         4.200    17.62659
#> 5   20-348.7-330  40660 0.9945355     3.971627         4.050    16.08509
#> 6   44-264.6-303  43064 0.7679523     4.054165         4.225    17.12885
#> 7     1-181-1249  40564 0.5376825     4.191180         4.200    17.60296
#> 8       2-76-690  43184 0.2506176     3.849929         3.825    14.72598
#> 9       2-76-691  40580 0.2509622     3.859651         3.825    14.76317
#> 10 14-161.1-1151 137772 0.4808628     7.043668         6.925    48.77740
#>    train_n positive_ob_n stixel_prevalence mean_non_zero_count i_stationary
#> 1     5471          1080        0.19740450            1.379630 0.0009782551
#> 2      788           186        0.23604060            1.354839 0.0009866625
#> 3     3429           920        0.26829980            1.604348 0.0102888000
#> 4     1752           389        0.22203200            2.272494 0.0040716770
#> 5     1069            96        0.08980355            1.072917 0.0000000000
#> 6     1469           358        0.24370320            1.254190 0.0087457020
#> 7      917            81        0.08833152            1.358025 0.0000000000
#> 8     2193           545        0.24851800            1.825688 0.0007947534
#> 9     1171           256        0.21861660            1.640625 0.0000000000
#> 10    2513           844        0.33585360            1.896919 0.0017567630
#>            year         day        time effort_hrs effort_distance_km
#> 1  0.0753970200 0.093119370 0.074023190 0.33446210        0.064391580
#> 2  0.0021844780 0.429846900 0.079498190 0.19003850        0.025708500
#> 3  0.0755529600 0.563472600 0.167291600 0.57488890        0.047441550
#> 4  0.0367294600 2.447557000 0.074426400 1.49242100        0.055355820
#> 5  0.0007547261 0.006725923 0.008760425 0.04499456        0.006688545
#> 6  0.0261479800 0.221674200 0.047343970 0.51563640        0.017106310
#> 7  0.0117553100 0.038932490 0.024324970 0.02840611        0.025159800
#> 8  0.0855764300 1.572839000 0.077741130 0.32078650        0.091529390
#> 9  0.0295483100 2.528617000 0.065845020 0.33829840        0.022476220
#> 10 0.0626012700 0.099072150 0.242724200 0.54061170        0.094723330
#>    number_observers      score        elev   northness     eastness
#> 1      0.0088865010 0.10977060 0.101280800 0.058624140 0.0498227900
#> 2      0.0055416540 0.25857680 0.038186440 0.008635983 0.0439463500
#> 3      0.0145845300 0.27700610 0.051606640 0.050119220 0.0807455400
#> 4      0.0106324500 0.14673300 0.024641730 0.043448080 0.0528379000
#> 5      0.0003075676 0.03260055 0.006023087 0.016163430 0.0037916700
#> 6      0.0047872150 0.18981830 0.039282690 0.033774600 0.0259320500
#> 7      0.0023204540 0.02805700 0.132365400 0.009956901 0.0009159445
#> 8      0.0059237920 0.15148180 0.027221590 0.026129700 0.0367949100
#> 9      0.0021075910 0.13850190 0.025894900 0.015056400 0.0407276700
#> 10     0.0056619270 0.39327290 0.429150600 0.138358900 0.1529488000
#>    umd_fs_c1_1500_pland umd_fs_c1_1500_lpi umd_fs_c1_1500_pd umd_fs_c1_1500_ed
#> 1          0.0027234820       0.0043938350      0.0058794630      1.090972e-02
#> 2          0.0003502530       0.0001266096      0.0004442638      1.582699e-03
#> 3          0.0060644240       0.0043841150      0.0082682620      7.763431e-03
#> 4          0.0003767716       0.0000000000      0.0015705880      0.000000e+00
#> 5          0.0001454525       0.0005360913      0.0001232081      1.034165e-04
#> 6          0.0011469630       0.0005356339      0.0031612130      7.653757e-04
#> 7          0.0009700515       0.0000000000      0.0029727750      5.057419e-06
#> 8          0.0030186030       0.0006602018      0.0018477210      1.618128e-02
#> 9          0.0001568978       0.0005858426      0.0022897110      6.114657e-04
#> 10         0.0017953200       0.0025331400      0.0123944400      8.705668e-03
#>    umd_fs_c2_1500_pland umd_fs_c2_1500_lpi umd_fs_c2_1500_pd umd_fs_c2_1500_ed
#> 1          6.536036e-05       1.385875e-04      1.026900e-03      1.232356e-03
#> 2          1.077418e-05       3.952960e-05      9.194824e-06      2.129982e-04
#> 3          5.425665e-05       5.851937e-04      6.861657e-04      2.957774e-04
#> 4          0.000000e+00       0.000000e+00      0.000000e+00      0.000000e+00
#> 5          1.635908e-04       5.725049e-03      1.998624e-04      3.461667e-04
#> 6          4.606220e-04       2.419310e-04      2.795407e-03      7.645322e-04
#> 7          0.000000e+00       1.295022e-10      0.000000e+00      2.368322e-06
#> 8          3.166677e-04       9.074745e-04      1.076650e-03      4.978517e-04
#> 9          7.054024e-06       7.327908e-05      3.181748e-04      1.245962e-04
#> 10         2.121127e-03       1.048519e-04      1.609226e-03      0.000000e+00
#>    umd_fs_c3_1500_pland umd_fs_c3_1500_lpi umd_fs_c3_1500_pd umd_fs_c3_1500_ed
#> 1          1.455139e-03       1.526986e-03      7.971544e-03      8.491477e-04
#> 2          0.000000e+00       3.034003e-05      0.000000e+00      2.807541e-02
#> 3          1.913617e-04       0.000000e+00      3.109620e-03      1.399894e-03
#> 4          4.595576e-05       1.206368e-04      0.000000e+00      8.711829e-04
#> 5          4.525723e-04       2.187336e-04      3.239615e-05      2.849426e-04
#> 6          8.663736e-05       5.011296e-04      4.581229e-04      3.872048e-04
#> 7          2.894102e-05       4.264299e-05      8.528113e-06      1.531512e-05
#> 8          9.545342e-04       1.052294e-03      1.098517e-02      1.097212e-03
#> 9          3.840394e-05       1.974018e-04      3.055801e-04      1.989313e-05
#> 10         1.718965e-03       8.333924e-04      1.390082e-02      2.397269e-03
#>    umd_fs_c4_1500_pland umd_fs_c4_1500_lpi umd_fs_c4_1500_pd umd_fs_c4_1500_ed
#> 1          0.0118347200       0.0322640700      0.0165489900      0.0099758480
#> 2          0.0003564676       0.0015225970      0.0002984091      0.0046034140
#> 3          0.0135092800       0.0180964800      0.0170723000      0.0194793600
#> 4          0.0389391300       0.0159553100      0.0000000000      0.0075254340
#> 5          0.0030635470       0.0003596819      0.0006753856      0.0002279311
#> 6          0.0033715010       0.0040507510      0.0034322360      0.0051245310
#> 7          0.0003196179       0.0095018490      0.0006281342      0.0010909900
#> 8          0.0000000000       0.0029066850      0.0036290150      0.0041386510
#> 9          0.0118694200       0.0014086050      0.0000000000      0.0070110180
#> 10         0.0227168000       0.0399350800      0.0257360200      0.0221969300
#>    umd_fs_c5_1500_pland umd_fs_c5_1500_lpi umd_fs_c5_1500_pd umd_fs_c5_1500_ed
#> 1          0.0643741700       0.0072241630      0.0094194450       0.007366266
#> 2          0.0011754040       0.0015831200      0.0002397078       0.005266960
#> 3          0.0263829500       0.0433506700      0.0531123600       0.025607830
#> 4          0.0122201100       0.0002632751      0.0000000000       0.003939228
#> 5          0.0026361000       0.0004159598      0.0000000000       0.001174661
#> 6          0.0008610701       0.0011230750      0.0197389700       0.001359950
#> 7          0.0628251900       0.0635573300      0.0020898600       0.001586942
#> 8          0.0025750870       0.0075747480      0.0041839780       0.012418150
#> 9          0.0015117300       0.0029289940      0.0130602500       0.004332934
#> 10         0.0396101900       0.0292842000      0.0312701700       0.042510870
#>    umd_fs_c6_1500_pland umd_fs_c6_1500_lpi umd_fs_c6_1500_pd umd_fs_c6_1500_ed
#> 1          5.233432e-05       1.380878e-04      5.841338e-03      6.391328e-04
#> 2          5.087715e-04       6.911137e-05      4.114848e-04      1.803258e-03
#> 3          1.954680e-04       1.408856e-03      3.875513e-03      7.496521e-04
#> 4          6.479944e-05       0.000000e+00      1.120315e-02      0.000000e+00
#> 5          1.379272e-04       8.474243e-04      2.499540e-03      4.439042e-04
#> 6          0.000000e+00       1.653208e-03      0.000000e+00      1.296503e-03
#> 7          6.777296e-06       0.000000e+00      2.346564e-06      2.330573e-07
#> 8          4.663898e-04       1.730161e-03      3.672262e-03      1.202988e-03
#> 9          1.093363e-03       1.679315e-04      0.000000e+00      5.898564e-05
#> 10         7.813094e-05       1.911201e-03      1.172361e-03      4.098645e-04
#>    umd_fs_c7_1500_pland umd_fs_c7_1500_lpi umd_fs_c7_1500_pd umd_fs_c7_1500_ed
#> 1          5.640498e-04       2.633142e-04      5.600501e-03      1.294010e-03
#> 2          2.755346e-04       1.431500e-04      4.869038e-04      0.000000e+00
#> 3          1.691197e-03       0.000000e+00      2.321885e-03      2.442502e-03
#> 4          3.858310e-05       3.448250e-04      6.560645e-04      1.270205e-03
#> 5          0.000000e+00       1.136329e-05      1.200172e-04      0.000000e+00
#> 6          5.652319e-05       3.417103e-04      1.953913e-03      7.240330e-04
#> 7          9.284303e-06       5.518037e-06      1.421568e-05      1.275092e-05
#> 8          1.768961e-04       6.690362e-04      2.805375e-04      7.285159e-04
#> 9          0.000000e+00       0.000000e+00      5.057841e-04      8.295934e-05
#> 10         6.142973e-04       5.773952e-04      3.394403e-03      2.398857e-03
#>    umd_fs_c8_1500_pland umd_fs_c8_1500_lpi umd_fs_c8_1500_pd umd_fs_c8_1500_ed
#> 1          2.472845e-03       1.000584e-02      0.0182494800      5.020096e-03
#> 2          6.402638e-04       1.630647e-03      0.0000000000      9.395646e-03
#> 3          4.805627e-03       3.913934e-03      0.0132522100      4.112604e-03
#> 4          5.550385e-03       1.201889e-03      0.0000000000      2.974373e-03
#> 5          9.689634e-05       5.511230e-05      0.0001275135      4.828050e-05
#> 6          2.874408e-03       2.599331e-03      0.0041826080      5.051250e-03
#> 7          0.000000e+00       3.419007e-05      0.0000000000      3.718155e-06
#> 8          1.159885e-03       3.923804e-03      0.0025455950      3.709961e-04
#> 9          0.000000e+00       2.302469e-03      0.0027952730      2.816046e-04
#> 10         2.270147e-03       1.535899e-02      0.0075878610      8.187648e-03
#>    umd_fs_c9_1500_pland umd_fs_c9_1500_lpi umd_fs_c9_1500_pd umd_fs_c9_1500_ed
#> 1          1.312570e-02       1.047461e-03      9.112189e-04      7.525897e-03
#> 2          5.082500e-05       1.792404e-05      4.504748e-05      6.439477e-06
#> 3          1.924896e-03       1.402168e-03      3.578811e-03      2.614111e-03
#> 4          1.124381e-02       3.329991e-05      2.852331e-03      3.470937e-04
#> 5          3.649386e-04       4.504763e-04      1.001990e-04      1.502924e-04
#> 6          1.660827e-03       1.152216e-03      6.901364e-04      3.474777e-04
#> 7          0.000000e+00       0.000000e+00      0.000000e+00      0.000000e+00
#> 8          0.000000e+00       1.424751e-03      8.798311e-05      9.476459e-04
#> 9          9.378029e-05       3.585386e-04      8.247225e-04      2.162118e-05
#> 10         1.543305e-03       1.677419e-03      1.318444e-03      7.325498e-03
#>    umd_fs_c10_1500_pland umd_fs_c10_1500_lpi umd_fs_c10_1500_pd
#> 1           5.077317e-03        0.0056275280       0.0054368170
#> 2           0.000000e+00        0.0000000000       0.0002489004
#> 3           4.310695e-03        0.0059748070       0.0131386800
#> 4           6.459465e-03        0.0055587130       0.0020258800
#> 5           3.654462e-04        0.0008241602       0.0010591500
#> 6           3.322055e-03        0.0016373330       0.0035058750
#> 7           4.802601e-05        0.0008022391       0.0010976530
#> 8           3.450309e-03        0.0045647170       0.0040236310
#> 9           6.876998e-04        0.0000000000       0.0086990240
#> 10          1.437191e-03        0.0081857350       0.0046592460
#>    umd_fs_c10_1500_ed umd_fs_c12_1500_pland umd_fs_c12_1500_lpi
#> 1        0.0057728050          0.0000000000        0.0130861700
#> 2        0.0009617416          0.0014104860        0.0000000000
#> 3        0.0049837320          0.0004563543        0.0063664180
#> 4        0.0044471880          0.0198884900        0.0152740600
#> 5        0.0014818100          0.0007562517        0.0022990800
#> 6        0.0041458280          0.0119007000        0.0205046500
#> 7        0.0020050760          0.0023693210        0.0007324885
#> 8        0.0034751460          0.0147784100        0.0186357900
#> 9        0.0016580520          0.0017732110        0.0003935473
#> 10       0.0071767080          0.0126661100        0.0201326300
#>    umd_fs_c12_1500_pd umd_fs_c12_1500_ed umd_fs_c13_1500_pland
#> 1        0.0139291800        0.004212341          3.146028e-03
#> 2        0.0046999840        0.005488440          0.000000e+00
#> 3        0.0151762200        0.031009010          1.256673e-02
#> 4        0.0117102800        0.018385990          4.098051e-02
#> 5        0.0006057489        0.001008587          3.747120e-05
#> 6        0.0164028700        0.011252710          2.225345e-02
#> 7        0.0017421320        0.003357973          1.481655e-05
#> 8        0.0158348000        0.027817910          9.872144e-03
#> 9        0.0048519320        0.005652433          3.904330e-03
#> 10       0.0381519400        0.024380870          4.836227e-04
#>    umd_fs_c13_1500_lpi umd_fs_c13_1500_pd umd_fs_c13_1500_ed
#> 1         1.548292e-02       6.761831e-03       0.0069343890
#> 2         1.775985e-03       1.018786e-03       0.0011464460
#> 3         9.735008e-03       2.293037e-03       0.0119386500
#> 4         1.261063e-02       1.010415e-02       0.0104906400
#> 5         0.000000e+00       4.095015e-03       0.0002266771
#> 6         1.662034e-02       5.295310e-03       0.0059594500
#> 7         1.623951e-05       5.204746e-07       0.0000226393
#> 8         4.312237e-02       7.012191e-03       0.0144625500
#> 9         1.780807e-01       7.079755e-03       0.0067691490
#> 10        9.201509e-04       1.488433e-03       0.0013423870
#>    umd_fs_c16_1500_pland umd_fs_c16_1500_lpi umd_fs_c16_1500_pd
#> 1           5.989389e-04        2.881679e-04       3.826464e-03
#> 2           5.475057e-05        2.568530e-05       2.000764e-05
#> 3           2.225828e-04        4.455246e-04       8.183452e-04
#> 4           1.863171e-04        2.164185e-04       3.619752e-04
#> 5           3.546646e-04        5.344158e-04       1.948389e-03
#> 6           3.526997e-04        5.015967e-05       0.000000e+00
#> 7           1.060825e-05        1.316949e-05       5.519794e-05
#> 8           5.724629e-04        1.358344e-04       2.238055e-03
#> 9           1.742886e-04        3.205752e-05       2.735311e-05
#> 10          2.211702e-03        1.160202e-03       0.000000e+00
#>    umd_fs_c16_1500_ed modiswater_fs_c0_1500_pland modiswater_fs_c0_1500_lpi
#> 1        9.059628e-04                           0                         0
#> 2        1.617079e-03                           0                         0
#> 3        3.279005e-03                           0                         0
#> 4        3.324844e-05                           0                         0
#> 5        2.763306e-04                           0                         0
#> 6        0.000000e+00                           0                         0
#> 7        1.291217e-06                           0                         0
#> 8        1.412964e-03                           0                         0
#> 9        0.000000e+00                           0                         0
#> 10       2.423434e-03                           0                         0
#>    modiswater_fs_c0_1500_pd modiswater_fs_c0_1500_ed
#> 1                         0                        0
#> 2                         0                        0
#> 3                         0                        0
#> 4                         0                        0
#> 5                         0                        0
#> 6                         0                        0
#> 7                         0                        0
#> 8                         0                        0
#> 9                         0                        0
#> 10                        0                        0
#>    modiswater_fs_c2_1500_pland modiswater_fs_c2_1500_lpi
#> 1                 1.067327e-03              2.581512e-03
#> 2                 1.905170e-02              0.000000e+00
#> 3                 5.693548e-03              8.845807e-03
#> 4                 0.000000e+00              0.000000e+00
#> 5                 1.495959e-05              0.000000e+00
#> 6                 2.764553e-03              5.532147e-03
#> 7                 9.599133e-07              4.296279e-06
#> 8                 0.000000e+00              0.000000e+00
#> 9                 6.526358e-04              3.118275e-06
#> 10                1.952885e-02              1.852254e-02
#>    modiswater_fs_c2_1500_pd modiswater_fs_c2_1500_ed
#> 1              0.000000e+00             4.556112e-03
#> 2              5.322000e-05             0.000000e+00
#> 3              6.550407e-03             9.315848e-03
#> 4              1.550674e-02             1.442754e-03
#> 5              2.865152e-05             1.672565e-06
#> 6              2.095630e-03             0.000000e+00
#> 7              0.000000e+00             2.975016e-06
#> 8              3.848252e-04             0.000000e+00
#> 9              0.000000e+00             0.000000e+00
#> 10             3.123477e-03             2.098731e-03
#>    modiswater_fs_c3_1500_pland modiswater_fs_c3_1500_lpi
#> 1                 4.066281e-03              3.121652e-03
#> 2                 9.150461e-04              2.019311e-03
#> 3                 6.144256e-03              1.295595e-02
#> 4                 0.000000e+00              7.203367e-02
#> 5                 1.127380e-04              8.890402e-03
#> 6                 1.002220e-03              1.348484e-04
#> 7                 8.165212e-06              1.111249e-06
#> 8                 1.270143e-03              2.783997e-03
#> 9                 1.656058e-04              5.654289e-04
#> 10                2.266942e-02              5.855875e-04
#>    modiswater_fs_c3_1500_pd modiswater_fs_c3_1500_ed
#> 1              7.289954e-04             4.689107e-03
#> 2              3.139606e-04             2.628940e-03
#> 3              7.743777e-03             9.265749e-03
#> 4              1.146467e-03             6.863937e-03
#> 5              2.665627e-04             2.400638e-05
#> 6              1.627095e-04             7.416909e-04
#> 7              8.293932e-07             4.544068e-06
#> 8              4.701189e-03             9.542242e-03
#> 9              4.057823e-03             1.067300e-03
#> 10             4.813483e-04             8.116956e-04
#>    modiswater_fs_c5_1500_pland modiswater_fs_c5_1500_lpi
#> 1                 0.0030293660              5.312359e-02
#> 2                 0.0000000000              0.000000e+00
#> 3                 0.0056210160              2.296938e-03
#> 4                 0.0005640737              2.279971e-04
#> 5                 0.0000000000              0.000000e+00
#> 6                 0.0040434390              3.756638e-03
#> 7                 0.0030735310              3.953502e-05
#> 8                 0.0000000000              0.000000e+00
#> 9                 0.0021510310              3.260836e-04
#> 10                0.0004609059              2.169040e-04
#>    modiswater_fs_c5_1500_pd modiswater_fs_c5_1500_ed
#> 1              7.567552e-03             7.557046e-04
#> 2              0.000000e+00             0.000000e+00
#> 3              8.548060e-03             1.216775e-03
#> 4              7.547949e-05             4.408866e-05
#> 5              0.000000e+00             0.000000e+00
#> 6              6.677018e-03             1.302557e-03
#> 7              1.684493e-05             7.273663e-07
#> 8              0.000000e+00             0.000000e+00
#> 9              4.722435e-05             8.709963e-05
#> 10             3.561498e-03             3.036658e-03
#>    modiswater_fs_c6_1500_pland modiswater_fs_c6_1500_lpi
#> 1                            0                         0
#> 2                            0                         0
#> 3                            0                         0
#> 4                            0                         0
#> 5                            0                         0
#> 6                            0                         0
#> 7                            0                         0
#> 8                            0                         0
#> 9                            0                         0
#> 10                           0                         0
#>    modiswater_fs_c6_1500_pd modiswater_fs_c6_1500_ed
#> 1                         0                        0
#> 2                         0                        0
#> 3                         0                        0
#> 4                         0                        0
#> 5                         0                        0
#> 6                         0                        0
#> 7                         0                        0
#> 8                         0                        0
#> 9                         0                        0
#> 10                        0                        0
#>    modiswater_fs_c7_1500_pland modiswater_fs_c7_1500_lpi
#> 1                            0                         0
#> 2                            0                         0
#> 3                            0                         0
#> 4                            0                         0
#> 5                            0                         0
#> 6                            0                         0
#> 7                            0                         0
#> 8                            0                         0
#> 9                            0                         0
#> 10                           0                         0
#>    modiswater_fs_c7_1500_pd modiswater_fs_c7_1500_ed       lon      lat
#> 1                         0                        0 -82.66239 43.16171
#> 2                         0                        0 -83.01648 43.71361
#> 3                         0                        0 -84.55903 44.98568
#> 4                         0                        0 -86.21622 41.79261
#> 5                         0                        0 -85.36707 42.01676
#> 6                         0                        0 -84.86500 42.85970
#> 7                         0                        0 -85.86240 42.68140
#> 8                         0                        0 -82.92381 42.43289
#> 9                         0                        0 -85.74791 43.50642
#> 10                        0                        0 -89.04060 46.70954
#>                          geometry
#> 1  POLYGON ((-86.71883 39.0992...
#> 2  POLYGON ((-85.07776 41.6511...
#> 3  POLYGON ((-87.541 42.09818,...
#> 4  POLYGON ((-88.31462 39.6926...
#> 5  POLYGON ((-87.35288 39.9917...
#> 6  POLYGON ((-86.89208 40.7472...
#> 7  POLYGON ((-87.95799 40.5814...
#> 8  POLYGON ((-84.84877 40.5203...
#> 9  POLYGON ((-87.67774 41.5939...
#> 10 POLYGON ((-92.56243 43.2470...