Converts cryptic cover class names to readable land cover names

Internal function that converts the cryptic predictor class names to readable land cover names.

convert_classes(cov_names, by_cover_class = FALSE, pretty = FALSE)

Arguments

cov_names
vector; Cover class names to convert.

by_cover_class
logical; Default is FALSE. If TRUE, replaces fragstat cover class name with a name for the cover class as whole.

pretty
logical; Default is FALSE. If TRUE, Converts from capital case to title case.

Value

A vector of converted names.

Examples

cns <- c("UMD_FS_C1_1500_PLAND", "MODISWATER_FS_C7_1500_LPI")

converted <- convert_classes(cov_names = cns, pretty = TRUE)