Deprecated functions in package ebirdst.

The functions listed below are deprecated and support for them will eventually be dropped. Help pages for deprecated functions are available at help("<function>-deprecated").

Usage

abundance_palette(
  n,
  season = c("weekly", "breeding", "nonbreeding", "migration", "prebreeding_migration",
    "postbreeding_migration", "year_round")
)

abundance_palette

For abundance_palette, use ebirdst_palettes