Label data cubes with the week date for each band

The data cubes are saved as GeoTIFFs, which don't allow for band labels. For convenience, this function labels the layers of a data cube once it has been loaded with the week dates for each band.

label_raster_stack(x)

Arguments

x
RasterStack or RasterBrick; original eBird Status and Trends data cube with 52 bands, one for each week.

Value

A RasterStack or RasterBrick with names assigned for the dates in the format of "wYYYY.MM.DD" per raster package constraints. The Raster* objects do not allow the names to start with a number, nor are they allowed to contain "-", so it is not possible to store the date in an ISO compliant format. Use parse_raster_dates() to convert the layer names to dates.

Examples

# download and load example abundance data
sp_path <- ebirdst_download("example_data")

#> Data already exists, use force = TRUE to re-download.

abd <- load_raster(sp_path, "abundance")

# label
abd <- label_raster_stack(abd)
names(abd)

#>  [1] "w2019.01.04" "w2019.01.11" "w2019.01.18" "w2019.01.25" "w2019.02.01"
#>  [6] "w2019.02.08" "w2019.02.15" "w2019.02.22" "w2019.03.01" "w2019.03.08"
#> [11] "w2019.03.15" "w2019.03.22" "w2019.03.29" "w2019.04.05" "w2019.04.12"
#> [16] "w2019.04.19" "w2019.04.26" "w2019.05.03" "w2019.05.10" "w2019.05.17"
#> [21] "w2019.05.24" "w2019.05.31" "w2019.06.07" "w2019.06.14" "w2019.06.21"
#> [26] "w2019.06.28" "w2019.07.06" "w2019.07.13" "w2019.07.20" "w2019.07.27"
#> [31] "w2019.08.03" "w2019.08.10" "w2019.08.17" "w2019.08.24" "w2019.08.31"
#> [36] "w2019.09.07" "w2019.09.14" "w2019.09.21" "w2019.09.28" "w2019.10.05"
#> [41] "w2019.10.12" "w2019.10.19" "w2019.10.26" "w2019.11.02" "w2019.11.09"
#> [46] "w2019.11.16" "w2019.11.23" "w2019.11.30" "w2019.12.07" "w2019.12.14"
#> [51] "w2019.12.21" "w2019.12.28"