Labels 52 week RasterStack with the dates for each band

The raster package does not allow layer names to be saved with the bands of a multi-band GeoTIFF. Accordingly, all eBird Status and Trends products raster results cover the entire 52 week temporal extent of analysis. For convenience, this function labels the RasterStack once it has been loaded with the dates for each band.

label_raster_stack(x)

Arguments

x
RasterStack or RasterBrick; original eBird Status and Trends product raster GeoTIFF with 52 bands, one for each week.

Value

A RasterStack or RasterBrick with names assigned for the dates in the format of "XYYYY.MM.DD" per raster package constraints. The Raster* objects do not allow the names to start with a number, nor are they allowed to contain "-", so it is not possible to store the date in an ISO compliant format.

Examples

# download and load example abundance data
sp_path <- ebirdst_download("example_data")

#> Data already exists, use force = TRUE to re-download.

abd <- load_raster("abundance", sp_path)

# label
abd <- label_raster_stack(abd)
names(abd)

#>  [1] "X2018.01.04" "X2018.01.11" "X2018.01.18" "X2018.01.25" "X2018.02.01"
#>  [6] "X2018.02.08" "X2018.02.15" "X2018.02.22" "X2018.03.01" "X2018.03.08"
#> [11] "X2018.03.15" "X2018.03.22" "X2018.03.29" "X2018.04.05" "X2018.04.12"
#> [16] "X2018.04.19" "X2018.04.26" "X2018.05.03" "X2018.05.10" "X2018.05.17"
#> [21] "X2018.05.24" "X2018.05.31" "X2018.06.07" "X2018.06.14" "X2018.06.21"
#> [26] "X2018.06.28" "X2018.07.06" "X2018.07.13" "X2018.07.20" "X2018.07.27"
#> [31] "X2018.08.03" "X2018.08.10" "X2018.08.17" "X2018.08.24" "X2018.08.31"
#> [36] "X2018.09.07" "X2018.09.14" "X2018.09.21" "X2018.09.28" "X2018.10.05"
#> [41] "X2018.10.12" "X2018.10.19" "X2018.10.26" "X2018.11.02" "X2018.11.09"
#> [46] "X2018.11.16" "X2018.11.23" "X2018.11.30" "X2018.12.07" "X2018.12.14"
#> [51] "X2018.12.21" "X2018.12.28"