Load full annual cycle map parameters

Get the map parameters used on the eBird Status and Trends website to optimally display the full annual cycle data. This includes bins for the abundance data, a projection, and an extent to map. The extent is the spatial extent of non-zero data across the full annual cycle and the projection is optimized for this extent.

load_fac_map_parameters(path)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

Value

A list containing elements:

custom_projection: a custom projection optimized for the given species' full annual cycle

fa_extent: an Extent object storing the spatial extent of non-zero data for the given species in the custom projection

fa_extent_sinu: the extent in sinusoidal projection

abundance_bins: abundance bins for the full annual cycle

Examples

# \donttest{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# get map parameters
load_fac_map_parameters(sp_path)

#> $custom_projection
#> [1] "+proj=laea +lat_0=40.8874715114569 +lon_0=-94.3562032160404"
#> 
#> $fa_extent
#> class      : Extent 
#> xmin       : -2470623 
#> xmax       : 2602377 
#> ymin       : -3469050 
#> ymax       : 3305950 
#> 
#> $fa_extent_sinu
#> class      : Extent 
#> xmin       : -11028917 
#> xmax       : -4522593 
#> ymin       : 941359.2 
#> ymax       : 7527684 
#> 
#> $abundance_bins
#>  [1] 0.01000000 0.01337109 0.02523799 0.04538258 0.05989215 0.07828108
#>  [7] 0.10139866 0.13024187 0.16597497 0.20995081 0.26373406 0.29485654
#> [13] 0.32912655 0.36681087 0.40819463 0.50329885 0.61712597 0.75272344
#> [19] 0.91353645 1.10344774 1.32682013 1.58854212 2.24951203 3.13790686
#> [25] 5.94350290
#> 

# }