Load full annual cycle map parameters

Get the map parameters used on the eBird Status and Trends website to optimally display the full annual cycle data. This includes bins for the abundance data, a projection, and an extent to map. The extent is the spatial extent of non-zero data across the full annual cycle and the projection is optimized for this extent.

load_fac_map_parameters(path)

Arguments

path
character; directory that the Status and Trends data for a given species was downloaded to. This path is returned by ebirdst_download() or get_species_path().

Value

A list containing elements:

custom_projection: a custom projection optimized for the given species' full annual cycle

fa_extent: an Extent object storing the spatial extent of non-zero data for the given species in the custom projection

res: a numeric vector with 2 elements giving the target resolution of raster in the custom projection.

fa_extent_sinu: the extent in sinusoidal projection

abundance_bins: abundance bins for the full annual cycle

Examples

# \donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
#> Data already exists, use force = TRUE to re-download.
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# get map parameters
load_fac_map_parameters(path)
#> $custom_projection
#> [1] "+proj=laea +lat_0=40.1815866905176 +lon_0=-90.9587386496754 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"
#> 
#> $fa_extent
#> class      : Extent 
#> xmin       : -2759160 
#> xmax       : 2769342 
#> ymin       : -3388622 
#> ymax       : 3258931 
#> 
#> $res
#> [1] 2958 2739
#> 
#> $fa_extent_sinu
#> class      : Extent 
#> xmin       : -11028917 
#> xmax       : -3959660 
#> ymin       : 991727 
#> ymax       : 7047709 
#> 
#> $abundance_bins
#>  [1] 9.603170e-06 1.015085e-01 6.599966e-01 7.293533e-01 7.851071e-01
#>  [6] 8.355833e-01 8.846025e-01 9.334762e-01 9.821025e-01 1.029958e+00
#> [11] 1.077345e+00 1.125083e+00 1.174886e+00 1.228218e+00 1.287000e+00
#> [16] 1.353333e+00 1.430667e+00 1.524504e+00 1.643603e+00 1.807917e+00
#> [21] 2.096790e+00 1.014804e+01
#> attr(,"labels")
#> [1] 0.1015085 1.1012143 2.0967896
#> 
# }