Load eBird Status and Trends raster data

Each of the eBird Status and Trends products is packaged as a GeoTIFF file with 52 bands, one for each week of the year. This function loads the data for a given product and species as a RasterStack object.

load_raster(
  product = c("abundance", "abundance_seasonal", "count", "occurrence",
    "abundance_lower", "abundance_upper", "template"),
  path
)

Arguments

product
character; status and trends product to load, options are relative abundance, seasonal abundance, count, occurrence, and upper and lower bounds on relative abundance. It is also possible to return a template raster with no data.

path
character; full path to the directory containing single species eBird Status and Trends products.

Value

A RasterStack with 52 layers for the given product, labelled by week. Seasonal abundance is the result of averaging the weekly abundance raster layers for each season or across the whole year for resident species. The date boundaries used for the seasonal definitions appear in ebirdst_runs and if a season failed review no associated layer will be included. There will be up to four layers labelled according to the seasons.

Details

The available raster layers are as follows:

occurrence: the expected probability of occurrence of the species, ranging from 0 to 1, on an eBird Traveling Count by a skilled eBirder starting at the optimal time of day with the optimal search duration and distance that maximizes detection of that species in a region.

count: the expected count of a species, conditional on its occurrence at the given location, on an eBird Traveling Count by a skilled eBirder starting at the optimal time of day with the optimal search duration and distance that maximizes detection of that species in a region.

abundance: the expected relative abundance, computed as the product of the probability of occurrence and the count conditional on occurrence, of the species on an eBird Traveling Count by a skilled eBirder starting at the optimal time of day with the optimal search duration and distance that maximizes detection of that species in a region.

abundance_seasonal: the expected relative abundance averaged across the weeks within each season.

abundance_lower: the lower 10th quantile of the expected relative abundance of the species on an eBird Traveling Count by a skilled eBirder starting at the optimal time of day with the optimal search duration and distance that maximizes detection of that species in a region.

abundance_upper: the upper 90th quantile of the expected relative abundance of the species on an eBird Traveling Count by a skilled eBirder starting at the optimal time of day with the optimal search duration and distance that maximizes detection of that species in a region.

Examples

# download example data
sp_path <- ebirdst_download("example_data")

#> Data already exists, use force = TRUE to re-download.

# load data
load_raster("abundance", sp_path)

#> class      : RasterStack 
#> dimensions : 243, 257, 62451, 52  (nrow, ncol, ncell, nlayers)
#> resolution : 2962.807, 2962.809  (x, y)
#> extent     : -7195045, -6433604, 4635982, 5355945  (xmin, xmax, ymin, ymax)
#> crs        : +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs 
#> names      : X2018.01.04, X2018.01.11, X2018.01.18, X2018.01.25, X2018.02.01, X2018.02.08, X2018.02.15, X2018.02.22, X2018.03.01, X2018.03.08, X2018.03.15, X2018.03.22, X2018.03.29, X2018.04.05, X2018.04.12, ... 
#> min values :   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.0000000,   0.2674534, ... 
#> max values :   0.3445658,   0.2800446,   0.2429540,   0.2598554,   0.2395974,   0.1694340,   0.1782648,   0.1106254,   0.1304015,   0.1717017,   0.3552348,   0.4426145,   0.5465971,   1.1958923,   2.2377937, ... 
#>