Stixel summary file loader

Internal function used by load_pis() to get the stixel summary information (from summary.txt).

load_summary(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing stixel summary information about each stixel centroid.

Examples

# \donttest{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# stixel summaries
summaries <- ebirdst:::load_summary(sp_path)
dplyr::glimpse(summaries)

#> Rows: 945
#> Columns: 24
#> $ stixel_id                <chr> "61-98.6-1205", "35-207.8-284", "99-166.8-32…
#> $ srd_n                    <int> 78104, 78252, 76112, 73404, 82496, 75540, 78…
#> $ lon                      <dbl> -84.76490, -83.29157, -84.34107, -84.39816, …
#> $ lat                      <dbl> 42.70233, 43.19564, 45.36555, 44.68508, 44.0…
#> $ date                     <dbl> 0.31286320, 0.61416530, 0.49918690, 0.173991…
#> $ stixel_width             <dbl> 4.69994, 5.99353, 5.73436, 5.84128, 4.64485,…
#> $ stixel_height            <dbl> 4.68954, 6.02181, 5.67542, 5.88858, 4.44974,…
#> $ stixel_area              <int> 171404, 171729, 167032, 161090, 181042, 1657…
#> $ train_n                  <int> 5000, 3977, 4629, 3768, 4437, 3808, 3855, 45…
#> $ positive_ob_n            <int> 1250, 227, 879, 18, 687, 58, 105, 759, 31, 1…
#> $ stixel_prevalence        <dbl> 0.250000000, 0.057078200, 0.189889800, 0.004…
#> $ mean_non_zero_count      <dbl> 1.558400, 1.405286, 1.732651, 1.111111, 1.72…
#> $ train_n_occ              <int> 5000, 3977, 4629, 3768, 4437, 3808, 3855, 45…
#> $ n_pos_occ                <int> 1250, 227, 879, 18, 687, 58, 105, 759, 31, 1…
#> $ train_n_abund            <int> 1792, 1414, 1738, 1281, 1654, 1353, 1385, 18…
#> $ n_pos_abund              <int> 1250, 1280, 1273, 1251, 1371, 1258, 1273, 15…
#> $ mean_occ_pred            <dbl> 0.3869277, 0.2180255, 0.3629538, 0.2881866, …
#> $ total_effort_hrs         <dbl> 5538.090, 4226.033, 5406.132, 2641.863, 5403…
#> $ total_effort_distance_km <dbl> 7151.708, 6050.446, 7920.236, 3677.514, 6903…
#> $ total_number_observers   <int> 8709, 6015, 8150, 5417, 7115, 5631, 5737, 88…
#> $ occ_thresh               <dbl> 0.38009700, 0.31261510, 0.34562460, 0.063066…
#> $ max_time                 <dbl> -6.740797, -7.738237, -6.870217, 2.311691, -…
#> $ max_effort_hrs           <dbl> 3.274174, 3.926348, 3.492304, 2.015522, 4.18…
#> $ max_effort_distance_km   <dbl> 4.726421, 8.421053, 9.473684, 1.052632, 9.47…

# }