Test data loader

Internal function used by compute_ppms() to get the test data for calculating predictive performance metrics. This file contains the observed counts and model predicted relative occurrence and abundance values.

load_test_data(path, return_sf = FALSE)

Arguments

path
character; full path to the directory containing single species eBird Status and Trends products.

return_sf
logical; whether to return an sf object of spatial points rather then the default data frame.

Value

data.frame containing test data, including checklist locations, observed counts, and predicted mean and associated error for relative occurrence and abundance.

Examples

# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# test data
test_data <- load_test_data(sp_path)
dplyr::glimpse(test_data)

#> Observations: 32,344
#> Variables: 15
#> $ row_id     <int> 5358763, 9356439, 5711862, 10565844, 2842301, 7184639, 754…
#> $ lon        <dbl> -83.72211, -85.63328, -86.18674, -85.76029, -82.61564, -84…
#> $ lat        <dbl> 42.28131, 44.75910, 42.78356, 41.84840, 42.65126, 42.26145…
#> $ date       <dbl> 0.21857920, 0.09562842, 0.36065570, 0.46174860, 0.57377050…
#> $ obs        <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ pi_mean    <dbl> 4.108974e-01, 3.482129e-04, 2.826249e-04, 1.977333e-03, NA…
#> $ pi_90      <dbl> 4.450487e-01, 6.940827e-04, 3.314300e-04, 2.434792e-03, NA…
#> $ pi_10      <dbl> 3.782367e-01, 1.068156e-04, 2.277767e-04, 1.539186e-03, NA…
#> $ pi_se      <dbl> 2.803735e-02, 2.334054e-04, 5.565813e-05, 4.272021e-04, NA…
#> $ pi_mu_mean <dbl> 5.492313e-01, 3.418820e-04, 2.967833e-04, 1.853338e-03, NA…
#> $ pi_mu_90   <dbl> 5.817485e-01, 6.996263e-04, 5.442580e-04, 2.604022e-03, NA…
#> $ pi_mu_10   <dbl> 5.122008e-01, 5.001256e-05, 1.605766e-04, 1.181451e-03, NA…
#> $ pi_mu_se   <dbl> 2.963667e-02, 2.514108e-04, 1.723854e-04, 7.132012e-04, NA…
#> $ pat        <dbl> 0.52000000, 0.00000000, 0.00000000, 0.00000000, 0.00000000…
#> $ pi_es      <int> 100, 48, 100, 74, 94, 99, 100, 98, 98, 100, 100, 100, 96, …