Parse data cube layer names into dates

label_raster_stack() labels the layers of a data cube with the associated week dates in the format of "wYYYY.MM.DD", because of constraints in the raster package. This function converts that character vector into an ISO compliant Date vector.

parse_raster_dates(x)

Arguments

x

Raster object; labeled Status and Trends data cube.

Value

Date vector.

Examples

if (FALSE) {
# download example data
path <- ebirdst_download("example_data")
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# weekly relative abundance
# note that only low resolution (lr) data are available for the example data
abd <- load_raster(path, "abundance", resolution = "lr")

# parse dates
parse_raster_dates(abd)
}