Plot partial dependency as line plot

For all of the available predictors in a single species set of eBird Status and Trends products, this function makes a line plot of a single partial dependency, with two options for smoothing.

plot_pds(pd_name, pds, st_extent, plot_quantiles = FALSE,
  pointwise_pi = TRUE, stixel_pds = FALSE, k.cont.res = 25,
  gbm.n.trees = 500, nnn.bs = 100, equivalent.ensemble.ss = 10,
  ci.alpha = 0.05, mean.all.data = FALSE, ylim = NA, print_plot = TRUE)

Arguments

pd_name
character; Single predictor name from PDs (via load_pds()). Unique predictors can be listed by calling pds <- load_pds(path) and then calling unique(pds$V4).

pds
data.frame; From load_pds().

st_extent
list; st_extent list for spatiotemporal filtering. Required, as results are less meaningful over large spatiotemporal extents.

plot_quantiles
logical; Default is FALSE. Adds a band for the upper (90th) and lower (10th) quantiles of the individual stixel PD values.

pointwise_pi
logical; Default is TRUE. A pointwise smoothing of individual stixel PD values. Ideal visualization of the PD values.

stixel_pds
logical; Default is FALSE. Toggle to plot the individual stixel PD values as semi-transparent lines.

k.cont.res
int; Default is 25. Number of knots to use in GAM based on continuining resolution of the current eBird Status and Trends products.

gbm.n.trees
int; Default is 500. Number of trees to use in pointwise GAM.

nnn.bs
int; Default is 100.

equivalent.ensemble.ss
int; Default is 10.

ci.alpha
numeric; Default is 0.05. Alpha transparency of confidence intervals.

mean.all.data
logical; Default is FALSE.

ylim
vector pair; Opportunity to pre-define plot y-min and y-max as vector pair (e.g., c(-1, 1)).

print_plot
logical; Default is TRUE. Set to FALSE to turn off plotting and only get return of pointwise pi values.

Value

Plots barplot and returns a list containing the quantiles (if plot_quantiles = TRUE) and/or the pointwise_pis upper, lower, and median (if pointwise_pi = TRUE).

Examples

# NOT RUN {
sp_path <- "path to species eBird Status and Trends products"
pds <- load_pds(sp_path)

ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

plot_pds(pd_name = "TIME", pds = pds, st_extent = ne_extent)
# }