Generate polygons for eBird Status and Trends stixels

eBird Status and Trends divides space and time into variably sized "stixels" within which individual base models are fit. The process of stixelization is performed many times and the prediction at any given point is the median of the predictions from all the stixels that that point falls in. load_stixels() loads information on all the stixels that compromise a species' Status and Trends model, with stixels identified by the location of their centroid. This function uses this information to define polygons for each stixel and attaches them to the original data in the form of an sf object.

stixelize(x)

Arguments

x

data.frame or sf object; stixel summary data loaded with load_stixels(), or any other data frame with fields lon, lat, stixel_width, and stixel_hight.

Value

sf object with geometry column storing polygons representing the stixels boundaries.

Examples

# \donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
#> Data already exists, use force = TRUE to re-download.
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# load stixel summary information
stixels <- load_stixels(path)

# build stixel polygons
stixelize(stixels)
#> Simple feature collection with 250 features and 96 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -92.44542 ymin: 40.66284 xmax: -79.71253 ymax: 50.67994
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>         stixel_id srd_n longitude latitude day_of_year stixel_width
#> 1   1-242.1-MR232 15788 -84.00009 44.91965   255.88623      2.57401
#> 2   1-185.8-OR142 15252 -83.57263 42.65483   199.24643      3.15333
#> 3    1-73.2-OR143 15196 -85.39118 42.74910    88.65478      3.16124
#> 4    1-45.0-OR142 14976 -86.46366 42.10683    59.27439      3.12097
#> 5   1-157.6-QP243 15540 -85.39507 44.05400   171.38546      3.00177
#> 6    2-78.9-OR231 15120 -84.41955 42.26159    94.65169      3.12982
#> 7   2-275.9-OR224 15512 -84.20342 43.79479   289.67532      2.72477
#> 8  11-320.3-SN121 15492 -83.24144 43.52412   334.50811      2.68200
#> 9  11-320.3-SN113 15036 -85.86891 42.22211   334.74232      2.50737
#> 10 11-320.3-SN112 15456 -85.55268 43.80916   333.97796      2.60381
#>    stixel_height stixel_area start_day_of_year end_day_of_year train_n
#> 1        1.78523      138591          242.8410        270.0000    1881
#> 2        2.21155      133886          186.0000        213.0000    3782
#> 3        2.21155      133394           73.8954        101.0000    4206
#> 4        2.18490      131463           45.0000         73.0000    3823
#> 5        2.07832      136414          158.0000        185.0000    4361
#> 6        2.18490      132727           79.0000        107.0000    4720
#> 7        1.89181      136168          276.0000        304.0000    2504
#> 8        1.86516      135993          321.0000        348.3954    7093
#> 9        1.78523      131990          321.0000        348.3954    3811
#> 10       1.81187      135677          321.0000        348.3954    4447
#>    positive_ob_n stixel_prevalence mean_non_zero_count train_n_occ n_pos_occ
#> 1            339      0.1802232855            1.383481        1881       339
#> 2             32      0.0084611317            1.656250        3782        32
#> 3            456      0.1084165478            1.609649        4206       456
#> 4             73      0.0190949516            1.136986        3823        73
#> 5            821      0.1882595735            1.533496        4361       821
#> 6            970      0.2055084746            1.659794        4720       970
#> 7            151      0.0603035144            1.370861        2504       151
#> 8              8      0.0011278726            1.250000           0         0
#> 9             61      0.0160062976            1.081967        3811        61
#> 10             1      0.0002248707            1.000000           0         0
#>    train_n_abund n_pos_abund is_assumed_zero neg_grid_cells pos_grid_cells
#> 1            864         654               0           1542            339
#> 2           1304        1255               0           3750             32
#> 3           1606        1350               0           3750            456
#> 4           1342        1254               0           3750             73
#> 5           2097        1640               0           3540            821
#> 6           2186        1788               0           3750            970
#> 7            903         792               0           2353            151
#> 8              0           0               1              0              0
#> 9           1349        1259               0           3750             61
#> 10             0           0               1              0              0
#>    stixel_n st_pos_n avg_spat_cov mean_occ_pred occ_thresh test_binary_tp
#> 1      3953      384    0.1017819     0.3848300 0.26782996             16
#> 2     16188       31    0.4109253     0.1589455 0.04589660              0
#> 3     22882      463    0.5539085     0.3015019 0.26569212             19
#> 4     20241       66    0.4100398     0.3453003 0.08799665              2
#> 5      8436      909    0.2404668     0.3687670 0.30913006             38
#> 6     31585     1210    0.6981503     0.3426001 0.34926806             51
#> 7      5803      151    0.1509541     0.2882827 0.21215288              4
#> 8      6484        8    0.1744981     0.0000000 0.00000000             NA
#> 9     12730       57    0.3240884     0.3242289 0.08175471              1
#> 10     4093        1    0.1114918     0.0000000 0.00000000             NA
#>    test_binary_fn test_binary_fp test_binary_tn test_binary_auc test_binary_pcc
#> 1              12             47            249       0.7063224       0.8179012
#> 2               0             11           1243              NA       0.9912281
#> 3               8            143           1478       0.8077433       0.9083738
#> 4               3             36           1484       0.6881579       0.9744262
#> 5              15             76            520       0.7947322       0.8597843
#> 6              13            290           1890       0.8319237       0.8649733
#> 7               6             30            416       0.6663677       0.9210526
#> 8              NA             NA             NA              NA              NA
#> 9               4             27            979       0.5865805       0.9693373
#> 10             NA             NA             NA              NA              NA
#>    test_binary_kappa test_binary_sensitivity test_binary_specificity
#> 1         0.26352288               0.5714286               0.8412162
#> 2         0.00000000                      NA               0.9912281
#> 3         0.17797069               0.7037037               0.9117829
#> 4         0.08773679               0.4000000               0.9763158
#> 5         0.38671222               0.7169811               0.8724832
#> 6         0.21410771               0.7968750               0.8669725
#> 7         0.15311597               0.4000000               0.9327354
#> 8                 NA                      NA                      NA
#> 9         0.05265544               0.2000000               0.9731610
#> 10                NA                      NA                      NA
#>    train_oob_binary_tp train_oob_binary_fn train_oob_binary_fp
#> 1                   12                   3                 128
#> 2                    0                   0                  51
#> 3                    4                   1                 470
#> 4                    0                   0                 145
#> 5                   43                  11                 199
#> 6                   43                   9                1011
#> 7                    0                   0                  79
#> 8                   NA                  NA                  NA
#> 9                    0                   0                  82
#> 10                  NA                  NA                  NA
#>    train_oob_binary_tn train_oob_binary_auc train_oob_binary_pcc
#> 1                  691            0.8218559            0.8429257
#> 2                 4606                   NA            0.9890487
#> 3                 6398            0.8657833            0.9314710
#> 4                 4769                   NA            0.9704925
#> 5                 1594            0.8426546            0.8863021
#> 6                 8275            0.8590248            0.8907689
#> 7                 1237                   NA            0.9399696
#> 8                   NA                   NA                   NA
#> 9                 3278                   NA            0.9755952
#> 10                  NA                   NA                   NA
#>    train_oob_binary_kappa train_oob_binary_sensitivity
#> 1              0.12645718                    0.8000000
#> 2              0.00000000                           NA
#> 3              0.01528369                    0.8000000
#> 4              0.00000000                           NA
#> 5              0.25492147                    0.7962963
#> 6              0.06786432                    0.8269231
#> 7              0.00000000                           NA
#> 8                      NA                           NA
#> 9              0.00000000                           NA
#> 10                     NA                           NA
#>    train_oob_binary_specificity test_occ_auc test_occ_pcc test_occ_kappa
#> 1                     0.8437118    0.6156915    0.6349206     0.05602606
#> 2                     0.9890487           NA           NA             NA
#> 3                     0.9315667    0.6455650    0.6481481     0.14229983
#> 4                     0.9704925           NA           NA             NA
#> 5                     0.8890128    0.7067175    0.7017544     0.32000000
#> 6                     0.8911264    0.7619337    0.5542522     0.16513111
#> 7                     0.9399696           NA           NA             NA
#> 8                            NA           NA           NA             NA
#> 9                     0.9755952           NA           NA             NA
#> 10                           NA           NA           NA             NA
#>    test_occ_bernoulli_dev test_occ_sensitivity test_occ_specificity
#> 1             -0.13199155            0.3125000            0.7446809
#> 2                      NA                   NA                   NA
#> 3             -0.94679975            0.6315789            0.6503497
#> 4                      NA                   NA                   NA
#> 5              0.03006469            0.5263158            0.7894737
#> 6             -0.96298034            0.8431373            0.5034483
#> 7                      NA                   NA                   NA
#> 8                      NA                   NA                   NA
#> 9                      NA                   NA                   NA
#> 10                     NA                   NA                   NA
#>    test_occ_mse test_occ_spearman train_oob_occ_auc train_oob_occ_pcc
#> 1     0.2248076         0.2086024         0.7526042         0.7642857
#> 2            NA                NA                NA                NA
#> 3     0.2477184         0.1667651                NA                NA
#> 4            NA                NA                NA                NA
#> 5     0.2131396         0.3632171         0.7324997         0.6652893
#> 6     0.2971317         0.3299120         0.8731627         0.5218216
#> 7            NA                NA                NA                NA
#> 8            NA                NA                NA                NA
#> 9            NA                NA                NA                NA
#> 10           NA                NA                NA                NA
#>    train_oob_occ_kappa train_oob_occ_bernoulli_dev train_oob_occ_sensitivity
#> 1           0.19512195                   -1.006697                 0.5833333
#> 2                   NA                          NA                        NA
#> 3                   NA                          NA                        NA
#> 4                   NA                          NA                        NA
#> 5           0.22772043                   -0.373959                 0.6744186
#> 6           0.06642063                   -3.875234                 0.9302326
#> 7                   NA                          NA                        NA
#> 8                   NA                          NA                        NA
#> 9                   NA                          NA                        NA
#> 10                  NA                          NA                        NA
#>    train_oob_occ_specificity train_oob_occ_mse train_oob_occ_spearman
#> 1                  0.7812500         0.1958517              0.2461886
#> 2                         NA                NA                     NA
#> 3                         NA                NA                     NA
#> 4                         NA                NA                     NA
#> 5                  0.6633166         0.2247569              0.3087839
#> 6                  0.5044510         0.3009381              0.2558968
#> 7                         NA                NA                     NA
#> 8                         NA                NA                     NA
#> 9                         NA                NA                     NA
#> 10                        NA                NA                     NA
#>    test_abd_spearman test_abd_poisson_dev test_log_abd_pearson test_abd_mad
#> 1          0.2632439           0.15862264            0.2999289    0.4983436
#> 2                 NA                   NA                   NA           NA
#> 3                 NA                   NA                   NA           NA
#> 4                 NA                   NA                   NA           NA
#> 5          0.3612416           0.12000672            0.4048512    0.6913270
#> 6          0.3494964           0.07073291            0.4606234    0.6072335
#> 7                 NA                   NA                   NA           NA
#> 8                 NA                   NA                   NA           NA
#> 9                 NA                   NA                   NA           NA
#> 10                NA                   NA                   NA           NA
#>    test_cnt_spearman test_log_cnt_pearson test_cnt_mad train_oob_abd_spearman
#> 1          0.5986653            0.5467011    0.5889206                     NA
#> 2                 NA                   NA           NA                     NA
#> 3                 NA                   NA           NA                     NA
#> 4                 NA                   NA           NA                     NA
#> 5          0.4686897            0.4575831    1.0997304                     NA
#> 6          0.4676870            0.5416575    0.8146352                     NA
#> 7                 NA                   NA           NA                     NA
#> 8                 NA                   NA           NA                     NA
#> 9                 NA                   NA           NA                     NA
#> 10                NA                   NA           NA                     NA
#>    train_oob_abd_poisson_dev train_oob_log_abd_pearson train_oob_abd_mad
#> 1                         NA                        NA                NA
#> 2                         NA                        NA                NA
#> 3                         NA                        NA                NA
#> 4                         NA                        NA                NA
#> 5                         NA                        NA                NA
#> 6                         NA                        NA                NA
#> 7                         NA                        NA                NA
#> 8                         NA                        NA                NA
#> 9                         NA                        NA                NA
#> 10                        NA                        NA                NA
#>    train_oob_cnt_spearman train_oob_log_cnt_pearson train_oob_cnt_mad
#> 1                      NA                        NA                NA
#> 2                      NA                        NA                NA
#> 3                      NA                        NA                NA
#> 4                      NA                        NA                NA
#> 5                      NA                        NA                NA
#> 6                      NA                        NA                NA
#> 7                      NA                        NA                NA
#> 8                      NA                        NA                NA
#> 9                      NA                        NA                NA
#> 10                     NA                        NA                NA
#>    total_effort_hrs total_effort_distance_km total_number_observers  max_time
#> 1          2340.806                 2768.663                   2675 -3.831527
#> 2          3864.727                 5588.941                   5234 -6.122889
#> 3          3484.124                 5052.824                   6389 -4.557751
#> 4          3121.110                 3888.280                   5718 -2.260556
#> 5          5116.226                 6983.482                   7636 -6.270952
#> 6          4437.322                 6456.359                   7041 -5.119362
#> 7          2640.959                 3505.578                   3303 -3.297684
#> 8             0.000                    0.000                      0        NA
#> 9          2960.313                 4011.860                   5852 -2.946500
#> 10            0.000                    0.000                      0        NA
#>    max_effort_hrs max_effort_distance_km    cds_u10     cds_v10    cds_d2m
#> 1        5.000000               9.880000 0.08935402  0.72378974 10.9128574
#> 2        2.483348              10.000000 0.87455226 -0.01465062 16.2915275
#> 3        2.500000               5.427087 1.66916824 -0.13360358  0.9571457
#> 4        6.000000               4.587000 0.73703615  0.41148152 -4.9815264
#> 5        3.250000               9.978000 0.54169251  0.25255864 12.8149007
#> 6        2.950000               4.876000 1.82092761  0.44051734  1.6549474
#> 7        3.750000               6.437000 0.94185865  1.04348265  8.8789348
#> 8              NA                     NA         NA          NA         NA
#> 9        2.333000               4.828000 1.69460609  0.58153352 -3.2201222
#> 10             NA                     NA         NA          NA         NA
#>       cds_t2m    cds_hcc cds_i10fg     cds_mcc     cds_lcc cds_sf       cds_rf
#> 1  17.0909184 0.00000000  6.192102 0.000000000 0.037172112      0 0.000000e+00
#> 2  24.4125343 0.12219798  6.353190 0.104725863 0.041917812      0 0.000000e+00
#> 3   8.7728453 0.14896312  8.422730 0.082584347 0.090488761      0 0.000000e+00
#> 4   0.4709681 0.07345917  7.494266 0.036592251 0.179710985      0 0.000000e+00
#> 5  19.2143289 0.16187264  6.132480 0.007355073 0.002105809      0 0.000000e+00
#> 6  10.6419797 0.27204614  8.857884 0.140249950 0.036958479      0 0.000000e+00
#> 7  14.3043691 0.04660248  7.788529 0.083286283 0.296171395      0 4.336809e-19
#> 8          NA         NA        NA          NA          NA     NA           NA
#> 9   1.6865247 0.00000000  7.549727 0.013367311 0.747806449      0 3.469447e-18
#> 10         NA         NA        NA          NA          NA     NA           NA
#>        cds_slc  cds_msl day_occ_slope  day_occ_mag day_cnt_slope day_cnt_mag
#> 1   -7.8677823 102151.4     34.417428 0.0219799377     37.704040  0.03067960
#> 2   -3.2616315 101525.0    590.876540 0.0004921375    -11.962687  0.08688800
#> 3  -13.5573852 101501.2      3.628831 0.1813758286      7.713641  0.08962059
#> 4  -12.6223468 102064.3     27.991563 0.0299109889     19.920614  0.03356456
#> 5   -0.9642756 101530.8     44.057629 0.0207078677      5.483490  0.20945861
#> 6   -6.7554705 101593.6      2.953618 0.2245522678      6.424392  0.10507233
#> 7  -12.6533979 101897.6     -5.251644 0.1498874947     22.723316  0.01400900
#> 8           NA       NA            NA           NA            NA          NA
#> 9   -6.5568797 102169.4     38.327767 0.0195140701    -23.445438  0.01654789
#> 10          NA       NA            NA           NA            NA          NA
#>                          geometry
#> 1  POLYGON ((-85.28709 44.0270...
#> 2  POLYGON ((-85.1493 41.54906...
#> 3  POLYGON ((-86.9718 41.64332...
#> 4  POLYGON ((-88.02414 41.0143...
#> 5  POLYGON ((-86.89596 43.0148...
#> 6  POLYGON ((-85.98446 41.1691...
#> 7  POLYGON ((-85.5658 42.84889...
#> 8  POLYGON ((-84.58244 42.5915...
#> 9  POLYGON ((-87.1226 41.32949...
#> 10 POLYGON ((-86.85458 42.9032...
# }