Generate polygons for eBird Status and Trends stixels

eBird Status and Trends divides space and time into variably sized "stixels" within which individual base models are fit. The process of stixelization is performed many times and the prediction at any given point is the median of the predictions from all the stixels that that point falls in. load_stixels() loads information on all the stixels that compromise a species' Status and Trends model, with stixels identified by the location of their centroid. This function uses this information to define polygons for each stixel and attaches them to the original data in the form of an sf object.

stixelize(x)

Arguments

x
data.frame or sf object; stixel summary data loaded with load_stixels(), or any other data frame with fields lon, lat, stixel_width, and stixel_hight.

Value

sf object with geometry column storing polygons representing the stixels boundaries.

Examples

# \donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
#> Data already exists, use force = TRUE to re-download.
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# load stixel summary information
stixels <- load_stixels(path)

# build stixel polygons
stixelize(stixels)
#> Simple feature collection with 3700 features and 29 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -93.86113 ymin: 40.56112 xmax: -78.80564 ymax: 50.21891
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>           stixel_id       lon      lat        date stixel_width stixel_height
#> 1  1-169.6-NSEWWEWW -86.72710 46.30995 0.500880326      4.73081       1.81187
#> 2  1-108.6-NSSEEWNW -83.54380 42.42114 0.344823682      3.23046       2.21154
#> 3  1-352.6-EWNWNSNN -82.83012 43.63808 0.005464481      3.33313       2.34478
#> 4  1-230.6-EWWNEWSE -84.03641 43.67917 0.673440100      3.43633       1.75858
#> 5  1-200.1-NSSWNENW -85.47172 44.18639 0.587080865      2.72994       2.23819
#> 6  1-230.6-EWWNEWSS -85.20056 41.99930 0.673173522      3.27365       1.75858
#> 7  1-291.6-NSWESSNN -84.42356 41.74085 0.837381337      4.27825       1.83852
#> 8   1-230.6-EWWNEWN -87.98485 46.48139 0.672797832      7.12135       3.57045
#> 9   1-78.1-NSSNEWWE -86.46522 44.04628 0.252486672      3.86278       1.70529
#> 10 1-200.1-NSSWNENS -82.90375 43.36570 0.587146268      2.71897       2.26483
#>    stixel_area train_n positive_ob_n stixel_prevalence mean_non_zero_count
#> 1        35201    2058           544        0.26433431            1.805147
#> 2        35095    4560           810        0.17763158            1.275309
#> 3        34902       0             0                NA                  NA
#> 4        35060    2700           181        0.06703704            1.337017
#> 5        35078    2620           294        0.11221374            1.591837
#> 6        35131    3766            16        0.00424854            1.187500
#> 7        35095    3835            85        0.02216428            1.105882
#> 8       140303    3562           432        0.12128018            1.550926
#> 9        35034    3539           458        0.12941509            1.908297
#> 10       35087    2181            83        0.03805594            1.409639
#>    train_n_occ n_pos_occ train_n_abund n_pos_abund mean_occ_pred
#> 1         2038       524           983         767     0.3933004
#> 2         4500       750          1738        1401     0.3313653
#> 3            0         0             0           0     0.0000000
#> 4         2687       168           978         860     0.2788611
#> 5         2609       283          1146         956     0.3259508
#> 6         3765        15          1295        1260     0.3598052
#> 7         3826        76          1363        1269     0.3254317
#> 8         3538       408          1290        1054     0.3381015
#> 9         3496       415          1473        1227     0.3173114
#> 10        2178        80           779         721     0.2215961
#>    total_effort_hrs total_effort_distance_km total_number_observers occ_thresh
#> 1          2093.094                 4344.600                   3306 0.38392039
#> 2          5585.933                 7301.025                   8485 0.31032601
#> 3             0.000                    0.000                      0 0.00000000
#> 4          3311.006                 4486.403                   3919 0.18839061
#> 5          3012.466                 3864.497                   3873 0.24341738
#> 6          3655.506                 4528.257                   5653 0.02355879
#> 7          2996.631                 3957.013                   5856 0.07706694
#> 8          4437.137                 6220.087                   5672 0.23287242
#> 9          3394.463                 4450.686                   5479 0.30808184
#> 10         2656.028                 3908.715                   2931 0.17881585
#>    is_assumed_zero neg_grid_cells pos_grid_cells stixel_n st_pos_n avg_spat_cov
#> 1                0           1514            527     3566      670   0.05311488
#> 2                0           3750            771    51468      892   0.35259644
#> 3                1              0              0     6323        5   0.08684635
#> 4                0           2519            180     6244      161   0.10274459
#> 5                0           2326            292     5515      283   0.07814688
#> 6                0           3750             16    11475       14   0.19109528
#> 7                0           3750             85     9789       75   0.09816320
#> 8                0           3130            423     7545      430   0.03286619
#> 9                0           3081            443     7941      420   0.11593919
#> 10               0           2098             83     4830       78   0.06775083
#>     max_time max_effort_hrs max_effort_distance_km
#> 1  -6.771908          3.000                  3.540
#> 2  -5.476377          3.000                  3.219
#> 3         NA             NA                     NA
#> 4  -6.419167          4.500                  6.759
#> 5  -6.227234          2.833                  6.437
#> 6  -5.591389          2.167                  3.219
#> 7  -3.865954          2.500                  3.219
#> 8  -5.987500          4.500                  7.242
#> 9  -5.758551          4.000                  2.414
#> 10 -6.613333          4.500                 10.000
#>                          geometry
#> 1  POLYGON ((-89.09251 45.4040...
#> 2  POLYGON ((-85.15903 41.3153...
#> 3  POLYGON ((-84.49668 42.4656...
#> 4  POLYGON ((-85.75457 42.7998...
#> 5  POLYGON ((-86.83669 43.0672...
#> 6  POLYGON ((-86.83738 41.1200...
#> 7  POLYGON ((-86.56269 40.8215...
#> 8  POLYGON ((-91.54552 44.6961...
#> 9  POLYGON ((-88.39661 43.1936...
#> 10 POLYGON ((-84.26323 42.2332...
# }