Generate polygons for eBird Status and Trends stixels

eBird Status and Trends divides space and time into variably sized "stixels" within which individual base models are fit. The process of stixelization is performed many times and the prediction at any given point is the median of the predictions from all the stixels that that point falls in. load_stixels() loads information on all the stixels that compromise a species' Status and Trends model, with stixels identified by the location of their centroid. This function uses this information to define polygons for each stixel and attaches them to the original data in the form of an sf object.

stixelize(x)

Arguments

x

data frame; stixel summary data loaded with load_stixels(), or any other data frame with fields lonitude_min, lontidue_max, latitude_min, and latitude_max.

Value

sf object with geometry column storing polygons representing the stixels boundaries.

Examples

if (FALSE) {
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# load stixel summary information
stixels <- load_stixels(path)

# build stixel polygons
stixelize(stixels)
}