Generate stixel polygons from PI data

All predictor importance data are provided at the stixel level. In these files, the stixel is defined based on a centroid, width, and height. This function uses this information to define polygons for each stixel and attaches them to the original data in the form of an sf object

stixelize(x)

# S3 method for data.frame
stixelize(x)

# S3 method for sf
stixelize(x)

Arguments

x
data.frame or sf object; PI data loaded with load_pis(), or any other data frame with fields lon, lat, stixel_width, and stixel_hight.

Value

sf object with geometry column storing polygons representing the stixels boundaries.

Methods (by class)

data.frame: PI or PD data

sf: PI or PD data as sf object

Examples

# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

#> Data already exists, use force = TRUE to re-download.

# load predictor importance
pis <- load_pis(sp_path)

stixelize(pis)

#> Simple feature collection with 1890 features and 83 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -92.95246 ymin: 38.64864 xmax: -79.32725 ymax: 51.0932
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>        stixel_id srd_n       lon      lat      date stixel_width stixel_height
#> 1   61-98.6-1205 78104 -84.76490 42.70233 0.3128632      4.69994       4.68954
#> 2   61-98.6-1205 78104 -84.76490 42.70233 0.3128632      4.69994       4.68954
#> 3   35-207.8-284 78252 -83.29157 43.19564 0.6141653      5.99353       6.02181
#> 4   35-207.8-284 78252 -83.29157 43.19564 0.6141653      5.99353       6.02181
#> 5   99-166.8-326 76112 -84.34107 45.36555 0.4991869      5.73436       5.67542
#> 6   99-166.8-326 76112 -84.34107 45.36555 0.4991869      5.73436       5.67542
#> 7    29-49.3-822 73404 -84.39816 44.68508 0.1739911      5.84128       5.88858
#> 8    29-49.3-822 73404 -84.39816 44.68508 0.1739911      5.84128       5.88858
#> 9  19-176.3-1247 82496 -84.84778 44.03590 0.5197299      4.64485       4.44974
#> 10 19-176.3-1247 82496 -84.84778 44.03590 0.5197299      4.64485       4.44974
#>    stixel_area train_n positive_ob_n stixel_prevalence mean_non_zero_count
#> 1       171404    5000          1250        0.25000000            1.558400
#> 2       171404    5000          1250        0.25000000            1.558400
#> 3       171729    3977           227        0.05707820            1.405286
#> 4       171729    3977           227        0.05707820            1.405286
#> 5       167032    4629           879        0.18988980            1.732651
#> 6       167032    4629           879        0.18988980            1.732651
#> 7       161090    3768            18        0.00477707            1.111111
#> 8       161090    3768            18        0.00477707            1.111111
#> 9       181042    4437           687        0.15483430            1.720524
#> 10      181042    4437           687        0.15483430            1.720524
#>    model i_stationary       year        day solar_noon_diff effort_hrs
#> 1    occ  24.18707000 10.1801800 78.9268100      40.2446700 129.823200
#> 2    abd  10.24263000 36.2267800 70.3328600     105.1565000  93.205910
#> 3    occ   0.38905790  1.1717870  1.3809290       8.9742190   8.905770
#> 4    abd   7.62567300 18.0336900 29.7783700      42.7801600  41.015810
#> 5    occ   3.54344200  9.0204600 17.8429800      30.9645900  66.917100
#> 6    abd  10.90930000 44.4792300 80.0185900     124.0631000 157.817400
#> 7    occ   0.05294615  0.6771610  0.1159236       0.2655264   1.036582
#> 8    abd   0.86835680  0.9879237  7.7037590       9.8005980  15.225370
#> 9    occ   1.17900700  4.5955350  8.2097440      18.0402600  34.650790
#> 10   abd  10.42443000 37.5335300 70.7147200      94.5145200  94.067850
#>    effort_distance_km number_observers       score eastness_median eastness_sd
#> 1          44.1411700       8.33171300  58.6079000      26.8404300  29.6457300
#> 2          89.0186400      27.37254000 131.6649000      61.0394200  50.9689000
#> 3           2.7944950       2.25546000   4.4803810       3.2779640   4.7464030
#> 4          49.6596800       8.18439400  34.8283800      25.9071500  24.6154600
#> 5          22.1657500       5.96007400  44.2075400      21.0891700  21.6299200
#> 6         159.9211000      26.38941000 119.8442000      75.1853700 112.6947000
#> 7           0.1231976       0.04656722   0.2142364       0.6255806   0.7512855
#> 8           9.8772990       2.95652300  13.3819900       5.4506480   3.7381450
#> 9           8.4439520       3.15378500  16.2911800      11.8642000  12.4472800
#> 10        105.6165000      26.95173000 145.6026000      54.0047600  70.7726500
#>    elev_median    elev_sd intertidal_fs_c1_1500_ed intertidal_fs_c1_1500_pland
#> 1   28.2940200 28.5690200                        0                           0
#> 2   74.6983800 81.1311200                        0                           0
#> 3    5.0742500  5.6214600                        0                           0
#> 4   24.6968800 54.0873500                        0                           0
#> 5   25.2590400 23.9697100                        0                           0
#> 6   76.1887300 70.8472800                        0                           0
#> 7    0.4947129  0.6661111                        0                           0
#> 8    1.8124090  3.1344200                        0                           0
#> 9   18.2243100 14.6229200                        0                           0
#> 10  61.1362800 88.3837300                        0                           0
#>    island mcd12q1_lccs1_fs_c1_1500_ed mcd12q1_lccs1_fs_c1_1500_pland
#> 1       0                  1.82164200                     1.25800700
#> 2       0                  6.92690000                     3.68125800
#> 3       0                  0.03674876                     0.05277983
#> 4       0                 11.65426000                    10.18063000
#> 5       0                  1.39911800                     0.99302300
#> 6       0                 15.58255000                    16.44564000
#> 7       0                  0.00000000                     0.00000000
#> 8       0                  0.00000000                     0.00000000
#> 9       0                  0.55352340                     0.70717020
#> 10      0                 19.31904000                    30.34296000
#>    mcd12q1_lccs1_fs_c2_1500_ed mcd12q1_lccs1_fs_c2_1500_pland
#> 1                            0                              0
#> 2                            0                              0
#> 3                            0                              0
#> 4                            0                              0
#> 5                            0                              0
#> 6                            0                              0
#> 7                            0                              0
#> 8                            0                              0
#> 9                            0                              0
#> 10                           0                              0
#>    mcd12q1_lccs1_fs_c11_1500_ed mcd12q1_lccs1_fs_c11_1500_pland
#> 1                     0.6585089                       0.8290236
#> 2                     0.6922234                       1.0719160
#> 3                     0.3121489                       0.3053405
#> 4                    26.1694300                      23.4049100
#> 5                     4.4127140                       3.5043990
#> 6                    17.6538200                      17.9486300
#> 7                     0.0000000                       0.0000000
#> 8                     0.0000000                       0.0000000
#> 9                     1.5799600                       1.0119940
#> 10                   29.0575800                      25.2682600
#>    mcd12q1_lccs1_fs_c12_1500_ed mcd12q1_lccs1_fs_c12_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c13_1500_ed mcd12q1_lccs1_fs_c13_1500_pland
#> 1                     0.5288506                     0.665394500
#> 2                     0.2907630                     0.346792200
#> 3                     0.1441503                     0.170333100
#> 4                    16.6392800                     9.859671000
#> 5                     0.2542677                     0.475688600
#> 6                    16.0169200                    15.599210000
#> 7                     0.0000000                     0.005142857
#> 8                     0.2312156                     0.603448000
#> 9                     0.4859604                     0.381008000
#> 10                   22.1463600                    36.349110000
#>    mcd12q1_lccs1_fs_c14_1500_ed mcd12q1_lccs1_fs_c14_1500_pland
#> 1                     21.996530                      20.0456500
#> 2                     44.898150                     114.6416000
#> 3                     14.786010                      10.1154400
#> 4                     26.924370                      21.1664300
#> 5                     28.371550                      33.8922800
#> 6                     78.860860                      90.6730600
#> 7                      0.232777                       0.1538259
#> 8                      1.614361                       1.7659350
#> 9                     26.269270                      30.9372300
#> 10                    63.530060                      76.1354600
#>    mcd12q1_lccs1_fs_c15_1500_ed mcd12q1_lccs1_fs_c15_1500_pland
#> 1                    8.01043400                       6.8004110
#> 2                   21.11212000                      18.5301000
#> 3                    2.65250600                       2.8771020
#> 4                   23.89627000                      40.7677400
#> 5                   12.29142000                       9.8089910
#> 6                   69.54450000                      72.2020700
#> 7                    0.03403406                       0.1198244
#> 8                    0.00000000                       0.0000000
#> 9                    6.34960700                       5.4850770
#> 10                 140.34770000                      55.4626300
#>    mcd12q1_lccs1_fs_c16_1500_ed mcd12q1_lccs1_fs_c16_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c21_1500_ed mcd12q1_lccs1_fs_c21_1500_pland
#> 1                     22.861860                      32.5837100
#> 2                     53.192940                      53.5446100
#> 3                      6.286148                      12.8456200
#> 4                     25.274860                      29.4880500
#> 5                     22.069560                      23.9928200
#> 6                     72.763710                      64.9640500
#> 7                      1.175361                       0.9157576
#> 8                      2.189837                       1.9130980
#> 9                     19.421980                      32.3080700
#> 10                    61.765020                      73.4411600
#>    mcd12q1_lccs1_fs_c22_1500_ed mcd12q1_lccs1_fs_c22_1500_pland
#> 1                    27.7817500                      30.9212800
#> 2                    42.6540700                      50.4776200
#> 3                     7.6620600                      14.2926100
#> 4                    17.0204600                      23.4018500
#> 5                    18.8724400                      17.1532500
#> 6                    37.6783300                      34.5785400
#> 7                     0.9907191                       0.8301785
#> 8                     4.7788170                       2.0469090
#> 9                    41.9058800                      66.1279300
#> 10                   43.4596500                      42.7917600
#>    mcd12q1_lccs1_fs_c255_1500_ed mcd12q1_lccs1_fs_c255_1500_pland
#> 1                              0                                0
#> 2                              0                                0
#> 3                              0                                0
#> 4                              0                                0
#> 5                              0                                0
#> 6                              0                                0
#> 7                              0                                0
#> 8                              0                                0
#> 9                              0                                0
#> 10                             0                                0
#>    mcd12q1_lccs1_fs_c31_1500_ed mcd12q1_lccs1_fs_c31_1500_pland
#> 1                    19.4717800                      25.3155000
#> 2                    43.1386300                      45.2810800
#> 3                     2.9219930                       7.0509290
#> 4                    24.6432400                      17.3307700
#> 5                    25.4481000                      23.6027600
#> 6                    31.6919400                      32.9481200
#> 7                     0.2881879                       0.6854411
#> 8                     0.9332024                       1.1621300
#> 9                    10.5408800                      16.3113400
#> 10                   46.2008300                      39.6852000
#>    mcd12q1_lccs1_fs_c32_1500_ed mcd12q1_lccs1_fs_c32_1500_pland
#> 1                    4.19176700                      4.44221000
#> 2                    9.70928400                      8.83644000
#> 3                    0.22879560                      0.79073050
#> 4                   12.58080000                     10.42160000
#> 5                    2.99687300                      2.48808500
#> 6                   16.20664000                     14.03066000
#> 7                    0.07516549                      0.03521078
#> 8                    0.43417560                      0.67279130
#> 9                    1.80027400                      1.08300100
#> 10                  19.92774000                     24.12544000
#>    mcd12q1_lccs1_fs_c41_1500_ed mcd12q1_lccs1_fs_c41_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c42_1500_ed mcd12q1_lccs1_fs_c42_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c43_1500_ed mcd12q1_lccs1_fs_c43_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs2_fs_c25_1500_ed mcd12q1_lccs2_fs_c25_1500_pland
#> 1                    11.8997800                      10.2390200
#> 2                    20.9668900                      21.2857100
#> 3                     1.1211720                       0.9209590
#> 4                    12.5088500                      13.0991700
#> 5                     3.3275580                       3.4273650
#> 6                    21.1664500                      22.5751300
#> 7                     0.2056323                       0.4047535
#> 8                     4.1069130                       1.9476850
#> 9                     4.9205120                       5.9711220
#> 10                   25.2547900                      21.9112900
#>    mcd12q1_lccs2_fs_c35_1500_ed mcd12q1_lccs2_fs_c35_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs2_fs_c36_1500_ed mcd12q1_lccs2_fs_c36_1500_pland
#> 1                    15.8521200                      30.3875300
#> 2                    25.5261600                      28.1239600
#> 3                     4.0767850                       5.3921200
#> 4                    23.9473300                      12.9201100
#> 5                     4.7243690                       5.1887840
#> 6                    28.4678600                      37.3769000
#> 7                     0.2511085                       0.3493021
#> 8                     0.9160119                       1.0508390
#> 9                     7.7205830                       9.7649460
#> 10                   29.7108400                      30.6608100
#>    mcd12q1_lccs3_fs_c27_1500_ed mcd12q1_lccs3_fs_c27_1500_pland
#> 1                    4.44595700                      3.51203700
#> 2                   17.41136000                     10.04045000
#> 3                    0.39134740                      0.46287210
#> 4                    9.50136900                     11.80231000
#> 5                    6.75439600                      4.75613000
#> 6                   34.21617000                     34.79181000
#> 7                    0.07111018                      0.01654412
#> 8                    0.00000000                      0.00000000
#> 9                    3.05751700                      3.05953200
#> 10                  48.37982000                     33.07469000
#>    mcd12q1_lccs3_fs_c50_1500_ed mcd12q1_lccs3_fs_c50_1500_pland
#> 1                    7.40833500                      6.82271500
#> 2                   24.30552000                     12.96813000
#> 3                    1.67022900                      2.08179300
#> 4                   11.06195000                     10.86274000
#> 5                    5.27820000                      3.58529000
#> 6                   15.37334000                     17.18862000
#> 7                    0.04681818                      0.01846154
#> 8                    0.49071630                      0.21196370
#> 9                    5.62996100                      3.24979200
#> 10                  27.54641000                     28.70575000
#>    mcd12q1_lccs3_fs_c51_1500_ed mcd12q1_lccs3_fs_c51_1500_pland
#> 1                    0.01705263                      0.01818182
#> 2                    0.00000000                      0.00000000
#> 3                    0.00000000                      0.00000000
#> 4                    0.00000000                      0.00000000
#> 5                    0.00000000                      0.00000000
#> 6                    0.00000000                      0.00000000
#> 7                    0.00000000                      0.00000000
#> 8                    0.00000000                      0.00000000
#> 9                    0.00000000                      0.00000000
#> 10                   0.00000000                      0.00000000
#>    mod44w_oic_fs_c1_1500_ed mod44w_oic_fs_c1_1500_pland
#> 1                         0                           0
#> 2                         0                           0
#> 3                         0                           0
#> 4                         0                           0
#> 5                         0                           0
#> 6                         0                           0
#> 7                         0                           0
#> 8                         0                           0
#> 9                         0                           0
#> 10                        0                           0
#>    mod44w_oic_fs_c2_1500_ed mod44w_oic_fs_c2_1500_pland
#> 1               15.46767000                 14.74223000
#> 2               33.64474000                 37.68181000
#> 3                1.34929300                  1.89335200
#> 4               24.77956000                 16.41344000
#> 5               12.33578000                 12.92125000
#> 6               62.44858000                 44.55207000
#> 7                0.03280719                  0.08289036
#> 8                1.15728200                  0.47897420
#> 9                6.16370900                  8.85562600
#> 10              52.55379000                 45.53346000
#>    mod44w_oic_fs_c3_1500_ed mod44w_oic_fs_c3_1500_pland northness_median
#> 1                         0                           0        25.433390
#> 2                         0                           0        57.628350
#> 3                         0                           0         3.573285
#> 4                         0                           0        23.137530
#> 5                         0                           0        21.966200
#> 6                         0                           0        97.390090
#> 7                         0                           0         1.069602
#> 8                         0                           0         3.939664
#> 9                         0                           0        11.679120
#> 10                        0                           0        72.930860
#>    northness_sd   ntl_mean     ntl_sd   occ_pred                       geometry
#> 1    26.0659000 40.4421200 38.0727900         NA POLYGON ((-87.11487 40.3575...
#> 2    66.8392900 53.8194000 66.4963200 487.638300 POLYGON ((-87.11487 40.3575...
#> 3     6.2836990 18.9002500 14.6203500         NA POLYGON ((-86.28833 40.1847...
#> 4    47.0065300 13.1281800 15.3372200  38.836020 POLYGON ((-86.28833 40.1847...
#> 5    23.2101800 30.6434000 25.3725400         NA POLYGON ((-87.20825 42.5278...
#> 6    93.6506500 48.1280900 61.5661300 490.692100 POLYGON ((-87.20825 42.5278...
#> 7     0.3094516  0.4001483  0.3258341         NA POLYGON ((-87.3188 41.74079...
#> 8     2.9094090  2.5334110  1.7408370   3.777097 POLYGON ((-87.3188 41.74079...
#> 9    11.7714700 43.0598200 43.0754200         NA POLYGON ((-87.1702 41.81103...
#> 10   79.8862600 58.1800900 42.0242000 256.427000 POLYGON ((-87.1702 41.81103...

# also works on sf objects
pis_sf <- sf::st_as_sf(pis, coords = c("lon", "lat"), crs = 4326)
stixelize(pis_sf)

#> Simple feature collection with 1890 features and 83 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -92.95246 ymin: 38.64864 xmax: -79.32725 ymax: 51.0932
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>        stixel_id srd_n      date stixel_width stixel_height stixel_area train_n
#> 1   61-98.6-1205 78104 0.3128632      4.69994       4.68954      171404    5000
#> 2   61-98.6-1205 78104 0.3128632      4.69994       4.68954      171404    5000
#> 3   35-207.8-284 78252 0.6141653      5.99353       6.02181      171729    3977
#> 4   35-207.8-284 78252 0.6141653      5.99353       6.02181      171729    3977
#> 5   99-166.8-326 76112 0.4991869      5.73436       5.67542      167032    4629
#> 6   99-166.8-326 76112 0.4991869      5.73436       5.67542      167032    4629
#> 7    29-49.3-822 73404 0.1739911      5.84128       5.88858      161090    3768
#> 8    29-49.3-822 73404 0.1739911      5.84128       5.88858      161090    3768
#> 9  19-176.3-1247 82496 0.5197299      4.64485       4.44974      181042    4437
#> 10 19-176.3-1247 82496 0.5197299      4.64485       4.44974      181042    4437
#>    positive_ob_n stixel_prevalence mean_non_zero_count model i_stationary
#> 1           1250        0.25000000            1.558400   occ  24.18707000
#> 2           1250        0.25000000            1.558400   abd  10.24263000
#> 3            227        0.05707820            1.405286   occ   0.38905790
#> 4            227        0.05707820            1.405286   abd   7.62567300
#> 5            879        0.18988980            1.732651   occ   3.54344200
#> 6            879        0.18988980            1.732651   abd  10.90930000
#> 7             18        0.00477707            1.111111   occ   0.05294615
#> 8             18        0.00477707            1.111111   abd   0.86835680
#> 9            687        0.15483430            1.720524   occ   1.17900700
#> 10           687        0.15483430            1.720524   abd  10.42443000
#>          year        day solar_noon_diff effort_hrs effort_distance_km
#> 1  10.1801800 78.9268100      40.2446700 129.823200         44.1411700
#> 2  36.2267800 70.3328600     105.1565000  93.205910         89.0186400
#> 3   1.1717870  1.3809290       8.9742190   8.905770          2.7944950
#> 4  18.0336900 29.7783700      42.7801600  41.015810         49.6596800
#> 5   9.0204600 17.8429800      30.9645900  66.917100         22.1657500
#> 6  44.4792300 80.0185900     124.0631000 157.817400        159.9211000
#> 7   0.6771610  0.1159236       0.2655264   1.036582          0.1231976
#> 8   0.9879237  7.7037590       9.8005980  15.225370          9.8772990
#> 9   4.5955350  8.2097440      18.0402600  34.650790          8.4439520
#> 10 37.5335300 70.7147200      94.5145200  94.067850        105.6165000
#>    number_observers       score eastness_median eastness_sd elev_median
#> 1        8.33171300  58.6079000      26.8404300  29.6457300  28.2940200
#> 2       27.37254000 131.6649000      61.0394200  50.9689000  74.6983800
#> 3        2.25546000   4.4803810       3.2779640   4.7464030   5.0742500
#> 4        8.18439400  34.8283800      25.9071500  24.6154600  24.6968800
#> 5        5.96007400  44.2075400      21.0891700  21.6299200  25.2590400
#> 6       26.38941000 119.8442000      75.1853700 112.6947000  76.1887300
#> 7        0.04656722   0.2142364       0.6255806   0.7512855   0.4947129
#> 8        2.95652300  13.3819900       5.4506480   3.7381450   1.8124090
#> 9        3.15378500  16.2911800      11.8642000  12.4472800  18.2243100
#> 10      26.95173000 145.6026000      54.0047600  70.7726500  61.1362800
#>       elev_sd intertidal_fs_c1_1500_ed intertidal_fs_c1_1500_pland island
#> 1  28.5690200                        0                           0      0
#> 2  81.1311200                        0                           0      0
#> 3   5.6214600                        0                           0      0
#> 4  54.0873500                        0                           0      0
#> 5  23.9697100                        0                           0      0
#> 6  70.8472800                        0                           0      0
#> 7   0.6661111                        0                           0      0
#> 8   3.1344200                        0                           0      0
#> 9  14.6229200                        0                           0      0
#> 10 88.3837300                        0                           0      0
#>    mcd12q1_lccs1_fs_c1_1500_ed mcd12q1_lccs1_fs_c1_1500_pland
#> 1                   1.82164200                     1.25800700
#> 2                   6.92690000                     3.68125800
#> 3                   0.03674876                     0.05277983
#> 4                  11.65426000                    10.18063000
#> 5                   1.39911800                     0.99302300
#> 6                  15.58255000                    16.44564000
#> 7                   0.00000000                     0.00000000
#> 8                   0.00000000                     0.00000000
#> 9                   0.55352340                     0.70717020
#> 10                 19.31904000                    30.34296000
#>    mcd12q1_lccs1_fs_c2_1500_ed mcd12q1_lccs1_fs_c2_1500_pland
#> 1                            0                              0
#> 2                            0                              0
#> 3                            0                              0
#> 4                            0                              0
#> 5                            0                              0
#> 6                            0                              0
#> 7                            0                              0
#> 8                            0                              0
#> 9                            0                              0
#> 10                           0                              0
#>    mcd12q1_lccs1_fs_c11_1500_ed mcd12q1_lccs1_fs_c11_1500_pland
#> 1                     0.6585089                       0.8290236
#> 2                     0.6922234                       1.0719160
#> 3                     0.3121489                       0.3053405
#> 4                    26.1694300                      23.4049100
#> 5                     4.4127140                       3.5043990
#> 6                    17.6538200                      17.9486300
#> 7                     0.0000000                       0.0000000
#> 8                     0.0000000                       0.0000000
#> 9                     1.5799600                       1.0119940
#> 10                   29.0575800                      25.2682600
#>    mcd12q1_lccs1_fs_c12_1500_ed mcd12q1_lccs1_fs_c12_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c13_1500_ed mcd12q1_lccs1_fs_c13_1500_pland
#> 1                     0.5288506                     0.665394500
#> 2                     0.2907630                     0.346792200
#> 3                     0.1441503                     0.170333100
#> 4                    16.6392800                     9.859671000
#> 5                     0.2542677                     0.475688600
#> 6                    16.0169200                    15.599210000
#> 7                     0.0000000                     0.005142857
#> 8                     0.2312156                     0.603448000
#> 9                     0.4859604                     0.381008000
#> 10                   22.1463600                    36.349110000
#>    mcd12q1_lccs1_fs_c14_1500_ed mcd12q1_lccs1_fs_c14_1500_pland
#> 1                     21.996530                      20.0456500
#> 2                     44.898150                     114.6416000
#> 3                     14.786010                      10.1154400
#> 4                     26.924370                      21.1664300
#> 5                     28.371550                      33.8922800
#> 6                     78.860860                      90.6730600
#> 7                      0.232777                       0.1538259
#> 8                      1.614361                       1.7659350
#> 9                     26.269270                      30.9372300
#> 10                    63.530060                      76.1354600
#>    mcd12q1_lccs1_fs_c15_1500_ed mcd12q1_lccs1_fs_c15_1500_pland
#> 1                    8.01043400                       6.8004110
#> 2                   21.11212000                      18.5301000
#> 3                    2.65250600                       2.8771020
#> 4                   23.89627000                      40.7677400
#> 5                   12.29142000                       9.8089910
#> 6                   69.54450000                      72.2020700
#> 7                    0.03403406                       0.1198244
#> 8                    0.00000000                       0.0000000
#> 9                    6.34960700                       5.4850770
#> 10                 140.34770000                      55.4626300
#>    mcd12q1_lccs1_fs_c16_1500_ed mcd12q1_lccs1_fs_c16_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c21_1500_ed mcd12q1_lccs1_fs_c21_1500_pland
#> 1                     22.861860                      32.5837100
#> 2                     53.192940                      53.5446100
#> 3                      6.286148                      12.8456200
#> 4                     25.274860                      29.4880500
#> 5                     22.069560                      23.9928200
#> 6                     72.763710                      64.9640500
#> 7                      1.175361                       0.9157576
#> 8                      2.189837                       1.9130980
#> 9                     19.421980                      32.3080700
#> 10                    61.765020                      73.4411600
#>    mcd12q1_lccs1_fs_c22_1500_ed mcd12q1_lccs1_fs_c22_1500_pland
#> 1                    27.7817500                      30.9212800
#> 2                    42.6540700                      50.4776200
#> 3                     7.6620600                      14.2926100
#> 4                    17.0204600                      23.4018500
#> 5                    18.8724400                      17.1532500
#> 6                    37.6783300                      34.5785400
#> 7                     0.9907191                       0.8301785
#> 8                     4.7788170                       2.0469090
#> 9                    41.9058800                      66.1279300
#> 10                   43.4596500                      42.7917600
#>    mcd12q1_lccs1_fs_c255_1500_ed mcd12q1_lccs1_fs_c255_1500_pland
#> 1                              0                                0
#> 2                              0                                0
#> 3                              0                                0
#> 4                              0                                0
#> 5                              0                                0
#> 6                              0                                0
#> 7                              0                                0
#> 8                              0                                0
#> 9                              0                                0
#> 10                             0                                0
#>    mcd12q1_lccs1_fs_c31_1500_ed mcd12q1_lccs1_fs_c31_1500_pland
#> 1                    19.4717800                      25.3155000
#> 2                    43.1386300                      45.2810800
#> 3                     2.9219930                       7.0509290
#> 4                    24.6432400                      17.3307700
#> 5                    25.4481000                      23.6027600
#> 6                    31.6919400                      32.9481200
#> 7                     0.2881879                       0.6854411
#> 8                     0.9332024                       1.1621300
#> 9                    10.5408800                      16.3113400
#> 10                   46.2008300                      39.6852000
#>    mcd12q1_lccs1_fs_c32_1500_ed mcd12q1_lccs1_fs_c32_1500_pland
#> 1                    4.19176700                      4.44221000
#> 2                    9.70928400                      8.83644000
#> 3                    0.22879560                      0.79073050
#> 4                   12.58080000                     10.42160000
#> 5                    2.99687300                      2.48808500
#> 6                   16.20664000                     14.03066000
#> 7                    0.07516549                      0.03521078
#> 8                    0.43417560                      0.67279130
#> 9                    1.80027400                      1.08300100
#> 10                  19.92774000                     24.12544000
#>    mcd12q1_lccs1_fs_c41_1500_ed mcd12q1_lccs1_fs_c41_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c42_1500_ed mcd12q1_lccs1_fs_c42_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs1_fs_c43_1500_ed mcd12q1_lccs1_fs_c43_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs2_fs_c25_1500_ed mcd12q1_lccs2_fs_c25_1500_pland
#> 1                    11.8997800                      10.2390200
#> 2                    20.9668900                      21.2857100
#> 3                     1.1211720                       0.9209590
#> 4                    12.5088500                      13.0991700
#> 5                     3.3275580                       3.4273650
#> 6                    21.1664500                      22.5751300
#> 7                     0.2056323                       0.4047535
#> 8                     4.1069130                       1.9476850
#> 9                     4.9205120                       5.9711220
#> 10                   25.2547900                      21.9112900
#>    mcd12q1_lccs2_fs_c35_1500_ed mcd12q1_lccs2_fs_c35_1500_pland
#> 1                             0                               0
#> 2                             0                               0
#> 3                             0                               0
#> 4                             0                               0
#> 5                             0                               0
#> 6                             0                               0
#> 7                             0                               0
#> 8                             0                               0
#> 9                             0                               0
#> 10                            0                               0
#>    mcd12q1_lccs2_fs_c36_1500_ed mcd12q1_lccs2_fs_c36_1500_pland
#> 1                    15.8521200                      30.3875300
#> 2                    25.5261600                      28.1239600
#> 3                     4.0767850                       5.3921200
#> 4                    23.9473300                      12.9201100
#> 5                     4.7243690                       5.1887840
#> 6                    28.4678600                      37.3769000
#> 7                     0.2511085                       0.3493021
#> 8                     0.9160119                       1.0508390
#> 9                     7.7205830                       9.7649460
#> 10                   29.7108400                      30.6608100
#>    mcd12q1_lccs3_fs_c27_1500_ed mcd12q1_lccs3_fs_c27_1500_pland
#> 1                    4.44595700                      3.51203700
#> 2                   17.41136000                     10.04045000
#> 3                    0.39134740                      0.46287210
#> 4                    9.50136900                     11.80231000
#> 5                    6.75439600                      4.75613000
#> 6                   34.21617000                     34.79181000
#> 7                    0.07111018                      0.01654412
#> 8                    0.00000000                      0.00000000
#> 9                    3.05751700                      3.05953200
#> 10                  48.37982000                     33.07469000
#>    mcd12q1_lccs3_fs_c50_1500_ed mcd12q1_lccs3_fs_c50_1500_pland
#> 1                    7.40833500                      6.82271500
#> 2                   24.30552000                     12.96813000
#> 3                    1.67022900                      2.08179300
#> 4                   11.06195000                     10.86274000
#> 5                    5.27820000                      3.58529000
#> 6                   15.37334000                     17.18862000
#> 7                    0.04681818                      0.01846154
#> 8                    0.49071630                      0.21196370
#> 9                    5.62996100                      3.24979200
#> 10                  27.54641000                     28.70575000
#>    mcd12q1_lccs3_fs_c51_1500_ed mcd12q1_lccs3_fs_c51_1500_pland
#> 1                    0.01705263                      0.01818182
#> 2                    0.00000000                      0.00000000
#> 3                    0.00000000                      0.00000000
#> 4                    0.00000000                      0.00000000
#> 5                    0.00000000                      0.00000000
#> 6                    0.00000000                      0.00000000
#> 7                    0.00000000                      0.00000000
#> 8                    0.00000000                      0.00000000
#> 9                    0.00000000                      0.00000000
#> 10                   0.00000000                      0.00000000
#>    mod44w_oic_fs_c1_1500_ed mod44w_oic_fs_c1_1500_pland
#> 1                         0                           0
#> 2                         0                           0
#> 3                         0                           0
#> 4                         0                           0
#> 5                         0                           0
#> 6                         0                           0
#> 7                         0                           0
#> 8                         0                           0
#> 9                         0                           0
#> 10                        0                           0
#>    mod44w_oic_fs_c2_1500_ed mod44w_oic_fs_c2_1500_pland
#> 1               15.46767000                 14.74223000
#> 2               33.64474000                 37.68181000
#> 3                1.34929300                  1.89335200
#> 4               24.77956000                 16.41344000
#> 5               12.33578000                 12.92125000
#> 6               62.44858000                 44.55207000
#> 7                0.03280719                  0.08289036
#> 8                1.15728200                  0.47897420
#> 9                6.16370900                  8.85562600
#> 10              52.55379000                 45.53346000
#>    mod44w_oic_fs_c3_1500_ed mod44w_oic_fs_c3_1500_pland northness_median
#> 1                         0                           0        25.433390
#> 2                         0                           0        57.628350
#> 3                         0                           0         3.573285
#> 4                         0                           0        23.137530
#> 5                         0                           0        21.966200
#> 6                         0                           0        97.390090
#> 7                         0                           0         1.069602
#> 8                         0                           0         3.939664
#> 9                         0                           0        11.679120
#> 10                        0                           0        72.930860
#>    northness_sd   ntl_mean     ntl_sd   occ_pred       lon      lat
#> 1    26.0659000 40.4421200 38.0727900         NA -84.76490 42.70233
#> 2    66.8392900 53.8194000 66.4963200 487.638300 -84.76490 42.70233
#> 3     6.2836990 18.9002500 14.6203500         NA -83.29157 43.19564
#> 4    47.0065300 13.1281800 15.3372200  38.836020 -83.29157 43.19564
#> 5    23.2101800 30.6434000 25.3725400         NA -84.34107 45.36555
#> 6    93.6506500 48.1280900 61.5661300 490.692100 -84.34107 45.36555
#> 7     0.3094516  0.4001483  0.3258341         NA -84.39816 44.68508
#> 8     2.9094090  2.5334110  1.7408370   3.777097 -84.39816 44.68508
#> 9    11.7714700 43.0598200 43.0754200         NA -84.84778 44.03590
#> 10   79.8862600 58.1800900 42.0242000 256.427000 -84.84778 44.03590
#>                          geometry
#> 1  POLYGON ((-87.11487 40.3575...
#> 2  POLYGON ((-87.11487 40.3575...
#> 3  POLYGON ((-86.28833 40.1847...
#> 4  POLYGON ((-86.28833 40.1847...
#> 5  POLYGON ((-87.20825 42.5278...
#> 6  POLYGON ((-87.20825 42.5278...
#> 7  POLYGON ((-87.3188 41.74079...
#> 8  POLYGON ((-87.3188 41.74079...
#> 9  POLYGON ((-87.1702 41.81103...
#> 10 POLYGON ((-87.1702 41.81103...